==> 1_exclusion.Rmd <==
---
title: "Exclusion"
output:
  html_document:
    toc: true
    toc_depth: 4
    toc_float: true
    code_folding: 'show'
    self_contained: false
---

# Reasons for Exclusion {.tabset}


The following reasons for exclusion apply:

* missing data (not finishing the initial survey)
* not female
* homosexuality
* currently in a homosexual relationship
* older than 50
* menopausal
* pregnant
* breastfeeding
* participants, who do not want to avoid pregnancy
* participants, who ‘took a chance’ on getting pregnant
* participants, who used no contraception for ‘other reasons’
* use of one of the following contraceptive methods: morning-after pill, breastfeeding, I am infertile, my partner is infertile, I am sterilized, my partner is sterilized, other
* incongruent information about current contraceptive method and method used in the last three months
* use of medication including sex hormones in the last three months.

## Data and Functions
```{r data and functions}
source("0_helpers.R")
load("data/cleaned.rdata")
# opts_chunk$set(warning = F, message = F, error = T)

library(knitr)
opts_chunk$set(fig.width = 9, fig.height = 7, cache = T, warning = F, message = F, cache = F, error = T)

# function to seperate the reasons for exclusion variable for upsetr
comma_separated_to_columns = function(df, col) {
  colname = deparse(substitute(col))
  df$splitcol = df %>% pull(colname)
  separate_rows(df, splitcol, convert = TRUE, sep = ", ") %>% 
    mutate(splitcol = if_else(is.na(splitcol), "no", 
                        if_else(splitcol == "" | 
                                  splitcol %in% c(), "included", as.character(splitcol)))) %>% 
    mutate(#splitcol = stringr::str_c(colname, "_", splitcol), 
           value = 1) %>% 
    spread(splitcol, value, fill = 0) %>% 
    select(-colname)
}

all_survey_length = nrow(all_surveys)
diary_length = nrow(diary)
diary_social_length = nrow(diary_social)
```

## Exclusion Steps {.tabset}
```{r reasons_for_exclusion}
all_surveys$reasons_for_exclusion = "" #insert new variable reasons_for_exclusion to list all exclusion criteria
all_surveys$reasons_for_exclusion_contraception = "" #insert new variable reasons_for_exclusion_contraception to list different contraceptions that lead to exclusion
```

### Missing data (not finishing the initial survey)
```{r didnt_finish_initial}
all_surveys = all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                       if_else(is.na(ended_initial), "didnt_finish_initial, ", "")))

kable(table(all_surveys$reasons_for_exclusion %contains% "didnt_finish_initial"))
```

### Not female
```{r not_female}
all_surveys = all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                       if_else(gender != 1 & !is.na(gender), "not_female, ", "")))

kable(table(all_surveys$reasons_for_exclusion %contains% "not_female"))
```

### Homosexuality
Not primarily heterosexual. This excludes women who reported being equally interested in men and women, women who reported being asexual or aromantic, and participants who did not identify as female gender.
```{r not_heterosexual_female}
all_surveys = all_surveys %>%
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                       if_else(sex_orientation >= 4 | gender != 1,
                                               "not_heterosexual_female, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion %contains% "not_heterosexual_female"))
```

### Currently in a non-heterosexual relationship
```{r}
all_surveys = all_surveys %>%
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                       if_else(partner_gender == 1,
                                               "non_heterosexual_relationship, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion %contains% "non_heterosexual_relationship"))
```

### Older than 50
```{r older_than_50}
all_surveys = all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(age >= 50, "older_than_50, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion %contains% "older_than_50"))
```

### Menopausal
```{r menopausal}
all_surveys = all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                       if_else(menopause_yes == 1 | menopause_yes == 2,
                                               "menopausal, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion %contains% "menopausal"))
```

### Pregnant
```{r pregnant}
all_surveys = all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                       if_else(pregnant == 1, "pregnant, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion %contains% "pregnant"))
```


### Breastfeeding
```{r breastfeeding}
all_surveys = all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(breast_feeding == 1, "breast_feeding, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion %contains% "breast_feeding"))
```

### Participants, who do not want to avoid pregnancy
```{r pregnant_trying}
all_surveys = all_surveys %>%
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                       if_else(pregnant_trying >= 4,
                                               "pregnant_trying, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion %contains% "pregnant_trying"))
```

### Participants, who ‘took a chance’ on getting pregnant
```{r}
all_surveys = all_surveys %>%
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                       if_else(contraception_at_all == 4,
                                               "pregnant_chance, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion %contains% "pregnant_chance"))
```

### Participants, who used no contraception for ‘other reasons’
```{r}
all_surveys = all_surveys %>%
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                       if_else(contraception_at_all == 6,
                                               "no_contraception, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion %contains% "no_contraception"))
```

### Conraceptive Method {.tabset}
Use of one of the following contraceptive methods


#### Morning-after pill
```{r contraception_hormonal_morning_after_pill}
all_surveys = all_surveys %>% 
  mutate(reasons_for_exclusion_contraception = str_c(
    reasons_for_exclusion_contraception,
    if_else(contraception_method %contains% "hormonal_morning_after_pill",
            "contraception_hormonal_morning_after_pill, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion_contraception %contains% "contraception_hormonal_morning_after_pill"))
```

#### Breastfeeding
```{r contraception_breast_feeding}
all_surveys = all_surveys %>% 
  mutate(reasons_for_exclusion_contraception = str_c(
    reasons_for_exclusion_contraception,
    if_else(contraception_method %contains% "breast_feeding",
            "contraception_breast_feeding, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion_contraception %contains% "contraception_breast_feeding"))
```

#### I am infertile
```{r contraception_infertile}
all_surveys = all_surveys %>% 
  mutate(reasons_for_exclusion_contraception = str_c(
    reasons_for_exclusion_contraception,
    if_else(contraception_method %contains% "infertile" &
              !(contraception_method %contains% "partner_infertile"),
            "contraception_infertile, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion_contraception %contains% "contraception_infertile"))
```

#### My partner is infertile
```{r contraception_partner_infertile}
all_surveys = all_surveys %>% 
  mutate(reasons_for_exclusion_contraception = str_c(
    reasons_for_exclusion_contraception,
    if_else(contraception_method %contains% "partner_infertile",
            "contraception_partner_infertile, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion_contraception %contains% "contraception_partner_infertile"))
```

#### I am sterilized
```{r contraception_sterilised}
all_surveys = all_surveys %>% 
  mutate(reasons_for_exclusion_contraception = str_c(
    reasons_for_exclusion_contraception,
    if_else(contraception_method %contains% "sterilised" &
              !(contraception_method %contains% "partner_sterilised"),
            "contraception_sterilised, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion_contraception %contains% "contraception_sterilised"))
```


#### My partner is sterilized
```{r contraception_partner_sterilised}
all_surveys = all_surveys %>% 
  mutate(reasons_for_exclusion_contraception = str_c(
    reasons_for_exclusion_contraception,
    if_else(contraception_method %contains% "partner_sterilised",
            "contraception_partner_sterilised, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion_contraception %contains% "contraception_partner_sterilised"))
```

#### Other
```{r contraception_not_listed}
all_surveys = all_surveys %>% 
  mutate(reasons_for_exclusion_contraception = str_c(
    reasons_for_exclusion_contraception,
    if_else(contraception_method %contains% "not_listed",
            "contraception_not_listed, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion_contraception %contains% "contraception_not_listed"))
```

#### Summary
All participants that will be excluded based on their choice of contraception
```{r}
all_surveys = all_surveys %>%
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                       if_else(reasons_for_exclusion_contraception != "",
                                       "contraceptive_method, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion %contains% "contraceptive_method"))
```



### Incongruent information about current contraceptive method and method used in the last three months
```{r incongruent_information_about_hormonal_contraception}
all_surveys = all_surveys %>%
  mutate(incongruent_information_about_hormonal_contraception =
           ifelse((hormonal_contraception_last3m == 0 & hormonal_contraception == T), 1, 0),
         reasons_for_exclusion = str_c(reasons_for_exclusion,
                                       if_else(incongruent_information_about_hormonal_contraception == 1,
                                               "incongruent_information_about_hormonal_contraception, ",
                                               "", "")))


kable(table(all_surveys$reasons_for_exclusion %contains% "incongruent_information_about_hormonal_contraception"))
```


### Use of medication including sex hormones in the last three months.
Taking sex hormones (other than the pill)
```{r}
all_surveys = all_surveys %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                       if_else(medication_name %contains% "Cycloprognova" |
                                                 medication_name %contains% "Cyproderm" |
                                                 medication_name %contains% "DHEA" |
                                                 medication_name %contains% "Hormone" |
                                                 medication_name %contains% "Cyclo-Progynova" |
                                                 medication_name %contains% "Femoston" |
                                                 medication_name %contains% "Gynokadin",
                                               "sex_hormones, ", "", "")))

kable(table(all_surveys$reasons_for_exclusion %contains% "sex_hormones"))
```

## Total number of excluded participants
```{r}
kable(table(all_surveys$reasons_for_exclusion == ""))
```
In total `r table(all_surveys$reasons_for_exclusion == "")["FALSE"]` participants were excluded, leaving `r table(all_surveys$reasons_for_exclusion == "")["TRUE"]` subjects which were included in the analysis.


## Reasons for exclusion {.tabset}
How to read this plot: The horizontal green bars show for how many women this reason for
exclusion applies. The blue bars show how many women are excluded for multiple reasons (e.g.,
they're menopausal _and_ not heterosexual).

### Table
```{r}
exclusion_reasons = all_surveys %>% 
  mutate(reasons_for_exclusion = str_sub(reasons_for_exclusion, 1, -3)) %>% 
  select(session, reasons_for_exclusion) %>% 
  comma_separated_to_columns(reasons_for_exclusion) %>% 
  select(-session) 

exclusion_reasons %>% 
  summarise_all(sum) %>%
  sort() %>% 
  gather(reason, n) %>% 
  left_join(all_surveys %>% mutate(reason = str_sub(reasons_for_exclusion, 1, -3)) %>%
              group_by(reason) %>%
              summarise(unique = n())) %>% 
  mutate(unique = if_else(is.na(unique), 0L, unique)) %>% 
  knitr::kable()
```

### Figure 3: Exclusion criteria and most common combinations of exclusion criteria. {.active}
The horizontal green bars show for how many women this exclusion criteria applies. The blue bars show how many women are excluded for multiple reasons (e.g., not finishing the initial survey and being pregnant). Only the 20 most common overlaps are displayed. See Figure S1 in the supplementary materials for all combinations.

```{r}
upset1 = exclusion_reasons %>% 
  rename(`Breastfeeding` = breast_feeding, 
         `Choice of Contraceptive Methode` = contraceptive_method,
         `Missing Data` = didnt_finish_initial,
         `Incongruent Information about Contraceptive Method` =
           incongruent_information_about_hormonal_contraception,
         `(Post-)Menopausal` = menopausal,
         `Using no Contraceptive Methods for Other Reasons` = no_contraception,
         `Not Female` = not_female,
         `Not Predominantly Heterosexual` = not_heterosexual_female,
         `Older than 50` = older_than_50,
         `Pregnant` = pregnant,
         `“Taking a Chance” of Becoming Pregnant` = pregnant_chance,
         `Trying to Become Pregnant` = pregnant_trying,
         `Medication Including Sex Hormones` = sex_hormones,
         `Currently in a Homosexual Relationship` = non_heterosexual_relationship) %>%
  filter(included == 0) %>% 
  select(-included) %>% 
  as.data.frame() %>%
  {
  upset(., ncol(.),
        20,
        order.by = "freq",
      main.bar.color = "#6E8691",
      matrix.color = "#6E8691",
      sets.bar.color = "#53AC9B",
      text.scale = 1.2)
  }
upset1

jpeg('Exclusion criteria and most common combinations of exclusion criteria.jpg', 
     width = 900, height = 470, quality = 1000)
upset1
dev.off()
```

### Figure S1: Exclusion criteria and all combinations of exclusion criteria. 
The horizontal green bars show for how many women this exclusion criteria applies. The blue bars show how many women are excluded for multiple reasons (e.g., not finishing the initial survey and being pregnant). All 52 combinations of exclusion criteria are displayed. 

```{r}
upset2 = exclusion_reasons %>% 
  rename(`Breastfeeding` = breast_feeding, 
         `Choice of Contraceptive Methode` = contraceptive_method,
         `Missing Data` = didnt_finish_initial,
         `Incongruent Information about Contraceptive Method` =
           incongruent_information_about_hormonal_contraception,
         `(Post-)Menopausal` = menopausal,
         `Using no Contraceptive Methods for Other Reasons` = no_contraception,
         `Not Female` = not_female,
         `Not Predominantly Heterosexual` = not_heterosexual_female,
         `Older than 50` = older_than_50,
         `Pregnant` = pregnant,
         `“Taking a Chance” of Becoming Pregnant` = pregnant_chance,
         `Trying to Become Pregnant` = pregnant_trying,
         `Medication Including Sex Hormones` = sex_hormones,
         `Currently in a Homosexual Relationship` = non_heterosexual_relationship) %>%
  filter(included == 0) %>% 
  select(-included) %>% 
  as.data.frame() %>% 
  {
  upset(., ncol(.), 80, order.by = "freq",
      main.bar.color = "#6E8691",
      matrix.color = "#6E8691",
      sets.bar.color = "#53AC9B",
      text.scale = 1.2)
  }
upset2

jpeg('Exclusion criteria and all combinations of exclusion criteria.jpg', 
     width = 900, height = 500, quality = 1000)
upset2
dev.off()
```



## Exlusion Steps Diary {.tabset}
To calculate Libido and Sexual Frequency we need Information from the diary. Therefore, two additional exlusion reasons for the diary apply:

* Skipping a diary day
* Dishonest answers on that day

```{r}
diary = diary %>% left_join(all_surveys %>% select(session, reasons_for_exclusion), by = 'session')
```


### Skipped this diary day (days after dropping out not included)
```{r}
diary = diary %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(is.na(ended_diary) & is.na(modified_diary),
                                                    "skipped_diary_entry, ", "")))

table(diary$reasons_for_exclusion %contains% "skipped_diary_entry")
```


### Disclosed that they responded dishonestly on that day.
```{r}
diary = diary %>% 
  mutate(reasons_for_exclusion = str_c(reasons_for_exclusion,
                                            if_else(dishonest_discard == 1,
                                                    "dishonest_answer, ", "", "")))

table(diary$reasons_for_exclusion %contains% "dishonest_answer")
```

## Create included variable
```{r}
all_surveys = all_surveys %>%
  mutate(included = if_else(reasons_for_exclusion == "", TRUE, FALSE))

diary = diary %>%
  mutate(included = if_else(reasons_for_exclusion == "", TRUE, FALSE))
```


## Save
```{r}
library(testthat)
expect_equal(nrow(diary), diary_length)
expect_equal(nrow(all_surveys), all_survey_length)

save(diary, all_surveys, file = "data/cleaned_selected.rdata")
```
