==> 3_descriptives.Rmd <==
---
title: "Descriptives"
output:
  html_document:
    toc: true
    toc_depth: 4
    toc_float: true
    code_folding: 'hide'
    self_contained: false
---


## Data and Functions
```{r data and function, results='hide',message=F,warning=F}
source("0_helpers.R")
knitr::opts_chunk$set(warning = FALSE, message = FALSE)
pander::panderOptions("table.split.table", Inf)
pander::panderOptions('round',2)
pander::panderOptions('digits',2)
pander::panderOptions('keep.trailing.zeros',TRUE)

load("data/cleaned_selected_wrangled.rdata")
```


## Means, Standard Deviations and Ranges
```{r}
mean_sd_range1 = data %>%
  select(session,
         age, education_years,
         bfi_extra, bfi_neuro, bfi_agree, bfi_consc, bfi_open,
         religiosity,
         diary_libido_mean, diary_masturbation_sum, diary_sex_active_sex_sum) %>%
  pivot_longer(-session, names_to = "Variable", values_to = "Value") %>%
  group_by(Variable) %>%
  summarise(n = sum(!is.na(Value)),
            mean = round(mean(Value, na.rm = T), 2),
            sd = round(sd(Value, na.rm = T), 2),
            min = round(min(Value, na.rm = T), 2),
            max = round(max(Value, na.rm = T), 2))


mean_sd_range2 = data %>%
  select(session,
         attractiveness_partner, 
         relationship_satisfaction, 
         satisfaction_sexual_intercourse,
         ) %>%
  pivot_longer(-session, names_to = "Variable", values_to = "Value") %>%
  group_by(Variable) %>%
  summarise(n = sum(!is.na(Value)),
            mean = round(mean(Value, na.rm = T), 2),
            sd = round(sd(Value, na.rm = T), 2),
            min = round(min(Value, na.rm = T), 2),
            max = round(max(Value, na.rm = T), 2))


mean_sd_range = data.frame(x = c(1:16)) %>%
  cbind(Variable = c("age", "education_years", "net_income", "bfi_extra", "bfi_neuro", "bfi_agree", "bfi_consc", "bfi_open", "religiosity", "relationship_duration", "attractiveness_partner", "relationship_satisfaction", "satisfaction_sexual_intercourse","diary_libido_mean", "diary_sex_active_sex_sum", "diary_masturbation_sum")) %>%
  select(-x)



mean_sd_range = left_join(mean_sd_range,
                          rbind(mean_sd_range1, mean_sd_range2),
                          by = "Variable")
                            
                            
                            
  
kable(mean_sd_range)
```

## Reliability {.tabset}
### Big Five Personality
```{r}
cronbachs_alpha_bfi_extra  = data %>%
  select(starts_with("bfi_extra_")) %>%
  psych::alpha()
cronbachs_alpha_bfi_extra

cronbachs_alpha_bfi_neuro  = data %>%
  select(starts_with("bfi_neuro_")) %>%
  psych::alpha()
cronbachs_alpha_bfi_neuro

cronbachs_alpha_bfi_agree  = data %>%
  select(starts_with("bfi_agree_")) %>%
  psych::alpha()
cronbachs_alpha_bfi_agree

cronbachs_alpha_bfi_consc  = data %>%
  select(starts_with("bfi_consc_")) %>%
  psych::alpha()
cronbachs_alpha_bfi_consc

cronbachs_alpha_bfi_open  = data %>%
  select(starts_with("bfi_open_")) %>%
  psych::alpha()
cronbachs_alpha_bfi_open

omega_bfi_extra  = data %>%
  select(starts_with("bfi_extra_")) %>%
  psych::omega()
omega_bfi_extra

omega_bfi_neuro  = data %>%
  select(starts_with("bfi_neuro_")) %>%
  psych::omega()
omega_bfi_neuro

omega_bfi_agree  = data %>%
  select(starts_with("bfi_agree_")) %>%
  psych::omega()
omega_bfi_agree

omega_bfi_consc  = data %>%
  select(starts_with("bfi_consc_")) %>%
  psych::omega()
omega_bfi_consc

omega_bfi_open  = data %>%
  select(starts_with("bfi_open_")) %>%
  psych::omega()
omega_bfi_open
```

### Attractiveness Partner
```{r}
cronbachs_alpha_attractiveness_partner  = data %>%
  select(starts_with("partner_attractiveness_")) %>%
  filter(!is.na(partner_attractiveness_body)) %>%
  psych::alpha()
cronbachs_alpha_attractiveness_partner
```


### Relationship Satisfaction
```{r}
cronbachs_alpha_relationship_satisfaction  = data %>%
  select(relationship_satisfaction_overall,
         relationship_satisfaction_2,
         relationship_satisfaction_3,
         relationship_problems_R,
         relationship_conflict_R) %>%
  filter(!is.na(relationship_satisfaction_overall)) %>%
  psych::alpha()
cronbachs_alpha_relationship_satisfaction

omega_relationship_satisfaction  = data %>%
  select(relationship_satisfaction_overall,
         relationship_satisfaction_2,
         relationship_satisfaction_3,
         relationship_problems_R,
         relationship_conflict_R) %>%
  filter(!is.na(relationship_satisfaction_overall)) %>%
  psych::omega()
omega_relationship_satisfaction
```

### Reliabilities {.active}
```{r}
reliability = data.frame(x = 1:7) %>%
  cbind(Variable = c("bfi_extra", "bfi_neuro", "bfi_agree", "bfi_consc", "bfi_open",
                  "attractiveness_partner", "relationship_satisfaction"),
        alpha = c(cronbachs_alpha_bfi_extra$total$std.alpha,
                  cronbachs_alpha_bfi_neuro$total$std.alpha,
                  cronbachs_alpha_bfi_agree$total$std.alpha,
                  cronbachs_alpha_bfi_consc$total$std.alpha,
                  cronbachs_alpha_bfi_open$total$std.alpha,
                  cronbachs_alpha_attractiveness_partner$total$std.alpha,
                  cronbachs_alpha_relationship_satisfaction$total$std.alpha),
        omega_h = c(omega_bfi_extra$omega_h,
                  omega_bfi_neuro$omega_h,
                  omega_bfi_agree$omega_h,
                  omega_bfi_consc$omega_h,
                  omega_bfi_open$omega_h,
                  NA,
                  omega_relationship_satisfaction$omega_h)) %>%
  mutate(alpha = round(alpha, 2),
         omega_h = round(omega_h, 2)) %>%
  select(-x)

kable(reliability)
  
```


## Summary {.active .tabset}
### Means, sds, ranges, and reliability estimeate
```{r}
summary = left_join(mean_sd_range, reliability, by = "Variable")
kable(summary)
```

### Zero-Order Correlations
```{r}
library(apaTables)

correlations = data %>%
  select(age, education_years,
         bfi_extra, bfi_neuro, bfi_agree, bfi_consc, bfi_open,
         religiosity,
         attractiveness_partner, 
         relationship_satisfaction, 
         satisfaction_sexual_intercourse,
         diary_libido_mean, diary_masturbation_sum, diary_sex_active_sex_sum)

correlations_table = apa.cor.table(correlations, filename = "Table.doc", table.number = 4)

correlations_table

```

### Means and sds by contraceptive group
```{r}
means_sd_congroup = data %>%
  group_by(contraception_hormonal, congruent_contraception) %>%
  summarize(count = n(),
            count_diary = sum(!is.na(diary_libido_mean)),
            age_mean = round(mean(age, na.rm = T), 2),
            age_sd = round(sd(age, na.rm = T), 2),
            education_years_mean = round(mean(education_years, na.rm = T), 2),
            education_years_sd = round(sd(education_years, na.rm = T), 2),
            bfi_extra_mean = round(mean(bfi_extra, na.rm = T), 2),
            bfi_extra_sd = round(sd(bfi_extra, na.rm = T), 2),
            bfi_neuro_mean = round(mean(bfi_neuro, na.rm = T), 2),
            bfi_neuro_sd = round(sd(bfi_neuro, na.rm = T), 2),
            bfi_agree_mean = round(mean(bfi_agree, na.rm = T), 2),
            bfi_agree_sd = round(sd(bfi_agree, na.rm = T), 2),
            bfi_consc_mean = round(mean(bfi_consc, na.rm = T), 2),
            bfi_consc_sd = round(sd(bfi_consc, na.rm = T), 2),
            bfi_open_mean = round(mean(bfi_open, na.rm = T), 2),
            bfi_open_sd = round(sd(bfi_open, na.rm = T), 2),
            religiosity_mean = round(mean(religiosity, na.rm = T), 2),
            religiosity_sd = round(sd(religiosity, na.rm = T), 2),
            attractiveness_partner_mean = round(mean(attractiveness_partner,
                                                     na.rm = T), 2),
            attractiveness_partner_sd = round(sd(attractiveness_partner,
                                                 na.rm = T), 2),
            relationship_satisfaction_mean = round(mean(relationship_satisfaction,
                                                     na.rm = T), 2),
            relationship_satisfaction_sd = round(sd(relationship_satisfaction,
                                                 na.rm = T), 2),
            satisfaction_sexual_intercourse_mean = round(
              mean(satisfaction_sexual_intercourse, na.rm = T), 2),
            satisfaction_sexual_intercourse_sd = round(
              sd(satisfaction_sexual_intercourse, na.rm = T), 2),
            diary_libido_mean_mean = round(mean(diary_libido_mean,
                                                     na.rm = T), 2),
            diary_libido_mean_sd = round(sd(diary_libido_mean,
                                                 na.rm = T), 2),
            diary_sex_active_sex_sum_mean = round(mean(diary_sex_active_sex_sum,
                                                     na.rm = T), 2),
            diary_sex_active_sex_sum_sd = round(sd(diary_sex_active_sex_sum,
                                                 na.rm = T), 2),
            diary_masturbation_sum_mean = round(mean(diary_masturbation_sum,
                                                     na.rm = T), 2),
            diary_masturbation_sum_sd = round(sd(diary_masturbation_sum,
                                                 na.rm = T), 2))

kable(means_sd_congroup)

crosstabs(~ relationship_duration_factor + congruent_contraception + contraception_hormonal,
          data = data)
```
