==> 4_codebook.Rmd <==
---
title: "Codebook Effects of Contraception"
output:
  html_document:
    toc: true
    toc_depth: 4
    toc_float: true
    code_folding: 'hide'
    self_contained: false
---
  
```{r results='hide',message=F,warning=F}
source("0_helpers.R")
knitr::opts_chunk$set(warning = FALSE, message = FALSE)
pander::panderOptions("table.split.table", Inf)
pander::panderOptions('round',2)
pander::panderOptions('digits',2)
pander::panderOptions('keep.trailing.zeros',TRUE)

load("data/cleaned_selected_wrangled.rdata")

knitr::opts_chunk$set(echo = FALSE, error = TRUE, dev = "CairoPNG")
library(ggplot2)
theme_set(theme_bw())

library(codebook)
```


## Selection Variables Codebook
```{r}
codebook_selection_data = data %>%
  select(-session,
         -starts_with("contraception_"),
         -starts_with("partner_"), -starts_with("relationship_satisfaction"),
         -relationship_problems_R, -relationship_conflict_R, -satisfaction_sexual_intercourse,
         -reasons_for_exclusion)

codebook(codebook_selection_data,  missingness_report = FALSE, indent = "##")
```

## Contraception Codebook
```{r}
codebook_contraception_data = data %>%
  select(contraception_hormonal, contraception_meeting_partner, contraception_change_since_meeting_partner)

codebook(codebook_contraception_data,  missingness_report = FALSE, indent = "##")
```

## Outcomes Codebook 
```{r}
codebook_outcome_data = data %>%
  select(partner_attractiveness_face,
         partner_attractiveness_body,
         relationship_satisfaction_overall,
         relationship_satisfaction_2,
         relationship_satisfaction_3,
         relationship_problems_R,
         relationship_conflict_R,
         satisfaction_sexual_intercourse,
         diary_libido_mean,
         diary_sex_active_mean,
         diary_sex_active_sex_mean)
codebook(codebook_outcome_data,  missingness_report = FALSE, indent = "##")
```


## Differences in contraception status by relationship status
```{r}
ggplot(data, aes(x = relationship_status, fill = contraception_hormonal)) + geom_bar(position = "dodge")
```

## Differences in Gaussian variables by HC and RS {.tabset}
### Hormonal contraception
```{r}
ggplot(data, aes(x = contraception_hormonal, y = partner_attractiveness_face)) + geom_jitter() + stat_summary(fun.y = mean, geom = "point", color = "red") + 
    stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red")

ggplot(data, aes(x = contraception_hormonal, y = partner_attractiveness_body)) + geom_jitter() + stat_summary(fun.y = mean, geom = "point", color = "red") + 
    stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red")

ggplot(data, aes(x = contraception_hormonal, y = relationship_satisfaction_overall)) + geom_jitter() + stat_summary(fun.y = mean, geom = "point", color = "red") + 
    stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red")

ggplot(data, aes(x = contraception_hormonal, y = relationship_satisfaction_2)) + geom_jitter() + stat_summary(fun.y = mean, geom = "point", color = "red") + 
    stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red")

ggplot(data, aes(x = contraception_hormonal, y = relationship_satisfaction_3)) + geom_jitter() + stat_summary(fun.y = mean, geom = "point", color = "red") + 
    stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red")

ggplot(data, aes(x = contraception_hormonal, y = relationship_problems_R)) + geom_jitter() + stat_summary(fun.y = mean, geom = "point", color = "red") + 
    stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red")

ggplot(data, aes(x = contraception_hormonal, y = relationship_conflict_R)) + geom_jitter() + stat_summary(fun.y = mean, geom = "point", color = "red") + 
    stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red")

ggplot(data, aes(x = contraception_hormonal, y = satisfaction_sexual_intercourse)) + geom_jitter() + stat_summary(fun.y = mean, geom = "point", color = "red") + 
    stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red")

ggplot(data, aes(x = contraception_hormonal, y = diary_libido_mean)) + geom_jitter() + stat_summary(fun.y = mean, geom = "point", color = "red") + 
    stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red")

ggplot(data, aes(x = contraception_hormonal, y = diary_sex_active_mean)) + geom_jitter() + stat_summary(fun.y = mean, geom = "point", color = "red") + 
    stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red")

ggplot(data, aes(x = contraception_hormonal, y = diary_sex_active_sex_mean)) + geom_jitter() + stat_summary(fun.y = mean, geom = "point", color = "red") + 
    stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red")
```

### Relationship status
```{r}
ggplot(data, aes(x = relationship_status, y = diary_libido_mean)) + geom_jitter() + stat_summary(fun.y = mean, geom = "point", color = "red") + 
    stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red")

ggplot(data, aes(x = relationship_status, y = diary_sex_active_mean)) + geom_jitter() + stat_summary(fun.y = mean, geom = "point", color = "red") + 
    stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red")

ggplot(data, aes(x = relationship_status, y = diary_sex_active_sex_mean)) + geom_jitter() + stat_summary(fun.y = mean, geom = "point", color = "red") + 
    stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red")
```


## Correlations {.tabset}
### Selection
```{r}
selection_variables = codebook_selection_data %>% select(
  -starts_with("bfi_agree_"), -starts_with("bfi_consc_"), -starts_with("bfi_open_"),
  -starts_with("bfi_extra_"), -starts_with("bfi_neuro_"),
  -starts_with("soi_r_attitude_"), -starts_with("soi_r_desire_"), -starts_with("soi_r_behavior_"), 
  -relationship_status, -net_income, -relationship_duration_factor, -relationship_duration,
  -number_of_days,
  -congruent_contraception)

colnames(selection_variables)
summary(selection_variables)
library(psych)
pairs.panels(selection_variables, cex.cor = 10)

cor_pred = cor(selection_variables, use = "pairwise.complete.obs")
print(cor_pred)
```

### Outcomes
```{r}
pairs.panels(codebook_outcome_data, cex.cor = 10)
cor_out = cor(codebook_outcome_data, use="complete.obs")
print(cor_out)
```
