==> 5_analyses_effects.Rmd <==
---
title: "Analyses Effects of Contraception"
output:
  html_document:
    toc: true
    toc_depth: 4
    toc_float: true
    code_folding: 'hide'
    self_contained: false
---
  
## Data 
```{r results='hide',message=F,warning=F}
source("0_helpers.R")

load("data/cleaned_selected_wrangled.rdata")
```
## Effects of Hormonal Contraceptives {.tabset}
### Attractiveness of Partner {.tabset}
#### Model
```{r}
m_hc_atrr = brm(attractiveness_partner ~ contraception_hormonal,
                data = data, family = gaussian(),
                file = "m_hc_atrr")
```

#### Summary
```{r}
kable(tidy(m_hc_atrr, intervals = T, prob = 0.90))
summary(m_hc_atrr, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_hc_atrr, range = c(-0.07, 0.07), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_hc_atrr, range = c(-0.07, 0.07), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_hc_atrr, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_hc_atrr %>%
  spread_draws(b_contraception_hormonalyes) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Contraception", NA),
         condition = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Hormonal Contraception", NA)) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.07))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.07, 0.07), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors") +
  xlim (-0.6, 0.6)
```

### Relationship Satisfaction {.tabset}
#### Model
```{r}
m_hc_relsat = brm(relationship_satisfaction ~ contraception_hormonal,
                data = data, family = gaussian(),
                file = "m_hc_relsat")
```

#### Summary
```{r}
kable(tidy(m_hc_relsat, intervals = T, prob = 0.90))
summary(m_hc_relsat, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_hc_relsat, range = c(-0.04, 0.04), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_hc_relsat, range = c(-0.04, 0.04), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_hc_relsat, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_hc_relsat %>%
  spread_draws(b_contraception_hormonalyes) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Contraception", NA),
         condition = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Hormonal Contraception", NA)) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.04))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.04, 0.04), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors") +
  xlim (-0.6, 0.6)
```


### Sexual Satisfaction {.tabset}
#### Model
```{r}
m_hc_sexsat = brm(satisfaction_sexual_intercourse ~ contraception_hormonal,
                data = data, family = gaussian(),
                file = "m_hc_sexsat")
```

#### Summary
```{r}
kable(tidy(m_hc_sexsat, intervals = T, prob = 0.90))
summary(m_hc_sexsat, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_hc_sexsat, range = c(-0.11, 0.11), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_hc_sexsat, range = c(-0.11, 0.11), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_hc_sexsat, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_hc_sexsat %>%
  spread_draws(b_contraception_hormonalyes) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Contraception", NA),
         condition = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Hormonal Contraception", NA)) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.11))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.11, 0.11), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors") +
  xlim (-0.6, 0.6)
```

### Libido

#### Model
```{r}
m_hc_libido = brm(diary_libido_mean ~ contraception_hormonal,
                data = data, family = gaussian(),
                file = "m_hc_libido")
```

#### Summary
```{r}
kable(tidy(m_hc_libido, intervals = T, prob = 0.90))
summary(m_hc_libido, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_hc_libido, range = c(-0.06, 0.06), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_hc_libido, range = c(-0.06, 0.06), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_hc_libido, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_hc_libido %>%
  spread_draws(b_contraception_hormonalyes) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Contraception", NA),
         condition = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Hormonal Contraception", NA)) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.06))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.06, 0.06), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors") +
  xlim (-0.6, 0.6)
```

### Sexual Frequency (Penetrative Intercourse) {.tabset}
#### Model
```{r}
m_hc_sexfreqpen = brm(diary_sex_active_sex_sum ~
                        offset(log(number_of_days)) +
                        contraception_hormonal,
                data = data, family = poisson(),
                file = "m_hc_sexfreqpen")
```

#### Summary
```{r}
kable(tidy(m_hc_sexfreqpen, intervals = T, prob = 0.90))
summary(m_hc_sexfreqpen, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_hc_sexfreqpen, range = c(-0.05, 0.05)), ci = 0.90,
                      parameters = "contraception")
equivalence_test(m_hc_sexfreqpen, range = c(-0.05, 0.05), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_hc_sexfreqpen,
                    effects = "contraception_hormonal",
                    conditions = data.frame(number_of_days = 1))
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_hc_sexfreqpen %>%
  spread_draws(b_contraception_hormonalyes) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Contraception", NA),
         condition = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Hormonal Contraception", NA)) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.05))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.05, 0.05), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```

### Masturbation Frequency {.tabset}
#### Model
```{r}
m_hc_masfreq = brm(diary_masturbation_sum ~ offset(log(number_of_days)) +
                     contraception_hormonal,
                data = data, family = poisson(),
                file = "m_hc_masfreq")
```

#### Summary
```{r}
kable(tidy(m_hc_masfreq, intervals = T, prob = 0.90))
summary(m_hc_masfreq, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_hc_masfreq, range = c(-0.05, 0.05), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_hc_masfreq, range = c(-0.05, 0.05), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_hc_masfreq,
                    effects = "contraception_hormonal",
                    conditions = data.frame(number_of_days = 1))
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_hc_masfreq %>%
  spread_draws(b_contraception_hormonalyes) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Contraception", NA),
         condition = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Hormonal Contraception", NA)) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.05))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.05, 0.05), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```

## Controlled Models: Effects of Hormonal Contraceptives {.tabset}
### Attractiveness of Partner {.tabset}
#### Model
```{r}
m_hc_atrr_controlled = brm(attractiveness_partner ~ contraception_hormonal +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
                data = data, family = gaussian(),
                file = "m_hc_atrr_controlled")
```

#### Summary
```{r}
kable(tidy(m_hc_atrr_controlled, intervals = T, prob = 0.90))
summary(m_hc_atrr_controlled, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_hc_atrr_controlled, range = c(-0.07, 0.07), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_hc_atrr_controlled, range = c(-0.07, 0.07), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_hc_atrr, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_hc_atrr_controlled %>%
  spread_draws(b_contraception_hormonalyes,
               b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes,
                        b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_relationship_duration_factor",
                        "Relationship Duration",
                        ifelse(condition %contains% "b_net_income",
                               "Income",
                               NA)),
         group = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Contraception", group),
         condition = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Hormonal Contraception", condition),
         condition = ifelse(condition == "b_age", "Age",
                ifelse(condition == "b_net_incomeeuro_500_1000", "500-1000 Euro",
                ifelse(condition == "b_net_incomeeuro_1000_2000", "1000-2000 Euro",
                ifelse(condition == "b_net_incomeeuro_2000_3000", "2000-3000 Euro",
                ifelse(condition == "b_net_incomeeuro_gt_3000", ">3000 Euro",
                ifelse(condition == "b_net_incomedont_tell", "do not tell",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto28months",
                       "13-28 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto52months",
                       "29-52 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_morethan52months",
                       ">52 months",
                ifelse(condition == "b_education_years", "Years of Education",
                ifelse(condition == "b_bfi_extra", "Extraversion",
                ifelse(condition == "b_bfi_neuro", "Neuroticism",
                ifelse(condition == "b_bfi_agree", "Agreeableness",
                ifelse(condition == "b_bfi_consc", "Conscientiousness",
                ifelse(condition == "b_bfi_open", "Openness",
                ifelse(condition == "b_religiosity", "Religiosity",
                       condition)))))))))))))))),
         group = ifelse(is.na(group), condition, group),
         condition = factor(condition, levels = rev(c("Hormonal Contraception", "Age",
                                        "500-1000 Euro", "1000-2000 Euro",
                                        "2000-3000 Euro", ">3000 Euro", "do not tell",
                                        "13-28 months", "29-52 months",
                                        ">52 months",
                                        "Years of Education",
                                        "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))),
         group = factor(group, levels = c("Contraception", "Age", "Income",
                                          "Relationship Duration","Years of Education",
                                          "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.07))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.07, 0.07), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```

### Relationship Satisfaction {.tabset}
#### Model
```{r}
m_hc_relsat_controlled = brm(relationship_satisfaction ~ contraception_hormonal +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
                data = data, family = gaussian(),
                file = "m_hc_relsat_controlled")
```

#### Summary
```{r}
kable(tidy(m_hc_relsat_controlled, intervals = T, prob = 0.90))
summary(m_hc_relsat_controlled, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_hc_relsat_controlled, range = c(-0.04, 0.04), ci = 0.90,
                      parameters = "contraception"))

equivalence_test(m_hc_relsat_controlled, range = c(-0.04, 0.04), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_hc_relsat_controlled, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_hc_relsat_controlled %>%
  spread_draws(b_contraception_hormonalyes,
               b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes,
                        b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_relationship_duration_factor",
                        "Relationship Duration",
                        ifelse(condition %contains% "b_net_income",
                               "Income",
                               NA)),
         group = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Contraception", group),
         condition = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Hormonal Contraception", condition),
         condition = ifelse(condition == "b_age", "Age",
                ifelse(condition == "b_net_incomeeuro_500_1000", "500-1000 Euro",
                ifelse(condition == "b_net_incomeeuro_1000_2000", "1000-2000 Euro",
                ifelse(condition == "b_net_incomeeuro_2000_3000", "2000-3000 Euro",
                ifelse(condition == "b_net_incomeeuro_gt_3000", ">3000 Euro",
                ifelse(condition == "b_net_incomedont_tell", "do not tell",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto28months",
                       "13-28 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto52months",
                       "29-52 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_morethan52months",
                       ">52 months",
                ifelse(condition == "b_education_years", "Years of Education",
                ifelse(condition == "b_bfi_extra", "Extraversion",
                ifelse(condition == "b_bfi_neuro", "Neuroticism",
                ifelse(condition == "b_bfi_agree", "Agreeableness",
                ifelse(condition == "b_bfi_consc", "Conscientiousness",
                ifelse(condition == "b_bfi_open", "Openness",
                ifelse(condition == "b_religiosity", "Religiosity",
                       condition)))))))))))))))),
         group = ifelse(is.na(group), condition, group),
         condition = factor(condition, levels = rev(c("Hormonal Contraception", "Age",
                                        "500-1000 Euro", "1000-2000 Euro",
                                        "2000-3000 Euro", ">3000 Euro", "do not tell",
                                        "13-28 months", "29-52 months",
                                        ">52 months",
                                        "Years of Education",
                                        "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))),
         group = factor(group, levels = c("Contraception", "Age", "Income",
                                          "Relationship Duration","Years of Education",
                                          "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.04))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.04, 0.04), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```


### Sexual Satisfaction {.tabset}
#### Model
```{r}
m_hc_sexsat_controlled = brm(satisfaction_sexual_intercourse ~ contraception_hormonal +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
                data = data, family = gaussian(),
                file = "m_hc_sexsat_controlled")
```

#### Summary
```{r}
kable(tidy(m_hc_sexsat_controlled, intervals = T, prob = 0.90))
summary(m_hc_sexsat_controlled, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_hc_sexsat_controlled, range = c(-0.11, 0.11), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_hc_sexsat_controlled, range = c(-0.11, 0.11), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_hc_sexsat_controlled, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_hc_sexsat_controlled %>%
  spread_draws(b_contraception_hormonalyes,
               b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes,
                        b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_relationship_duration_factor",
                        "Relationship Duration",
                        ifelse(condition %contains% "b_net_income",
                               "Income",
                               NA)),
         group = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Contraception", group),
         condition = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Hormonal Contraception", condition),
         condition = ifelse(condition == "b_age", "Age",
                ifelse(condition == "b_net_incomeeuro_500_1000", "500-1000 Euro",
                ifelse(condition == "b_net_incomeeuro_1000_2000", "1000-2000 Euro",
                ifelse(condition == "b_net_incomeeuro_2000_3000", "2000-3000 Euro",
                ifelse(condition == "b_net_incomeeuro_gt_3000", ">3000 Euro",
                ifelse(condition == "b_net_incomedont_tell", "do not tell",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto28months",
                       "13-28 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto52months",
                       "29-52 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_morethan52months",
                       ">52 months",
                ifelse(condition == "b_education_years", "Years of Education",
                ifelse(condition == "b_bfi_extra", "Extraversion",
                ifelse(condition == "b_bfi_neuro", "Neuroticism",
                ifelse(condition == "b_bfi_agree", "Agreeableness",
                ifelse(condition == "b_bfi_consc", "Conscientiousness",
                ifelse(condition == "b_bfi_open", "Openness",
                ifelse(condition == "b_religiosity", "Religiosity",
                       condition)))))))))))))))),
         group = ifelse(is.na(group), condition, group),
         condition = factor(condition, levels = rev(c("Hormonal Contraception", "Age",
                                        "500-1000 Euro", "1000-2000 Euro",
                                        "2000-3000 Euro", ">3000 Euro", "do not tell",
                                        "13-28 months", "29-52 months",
                                        ">52 months",
                                        "Years of Education",
                                        "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))),
         group = factor(group, levels = c("Contraception", "Age", "Income",
                                          "Relationship Duration","Years of Education",
                                          "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.11))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.11, 0.11), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```

### Libido

#### Model
```{r}
m_hc_libido_controlled = brm(diary_libido_mean ~ contraception_hormonal +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
                data = data, family = gaussian(),
                file = "m_hc_libido_controlled")
```

#### Summary
```{r}
kable(tidy(m_hc_libido_controlled, intervals = T, prob = 0.90))
summary(m_hc_libido_controlled, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_hc_libido_controlled, range = c(-0.06, 0.06), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_hc_libido_controlled, range = c(-0.06, 0.06), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_hc_libido_controlled, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_hc_libido_controlled %>%
  spread_draws(b_contraception_hormonalyes,
               b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes,
                        b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_relationship_duration_factor",
                        "Relationship Duration",
                        ifelse(condition %contains% "b_net_income",
                               "Income",
                               NA)),
         group = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Contraception", group),
         condition = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Hormonal Contraception", condition),
         condition = ifelse(condition == "b_age", "Age",
                ifelse(condition == "b_net_incomeeuro_500_1000", "500-1000 Euro",
                ifelse(condition == "b_net_incomeeuro_1000_2000", "1000-2000 Euro",
                ifelse(condition == "b_net_incomeeuro_2000_3000", "2000-3000 Euro",
                ifelse(condition == "b_net_incomeeuro_gt_3000", ">3000 Euro",
                ifelse(condition == "b_net_incomedont_tell", "do not tell",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto28months",
                       "13-28 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto52months",
                       "29-52 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_morethan52months",
                       ">52 months",
                ifelse(condition == "b_education_years", "Years of Education",
                ifelse(condition == "b_bfi_extra", "Extraversion",
                ifelse(condition == "b_bfi_neuro", "Neuroticism",
                ifelse(condition == "b_bfi_agree", "Agreeableness",
                ifelse(condition == "b_bfi_consc", "Conscientiousness",
                ifelse(condition == "b_bfi_open", "Openness",
                ifelse(condition == "b_religiosity", "Religiosity",
                       condition)))))))))))))))),
         group = ifelse(is.na(group), condition, group),
         condition = factor(condition, levels = rev(c("Hormonal Contraception", "Age",
                                        "500-1000 Euro", "1000-2000 Euro",
                                        "2000-3000 Euro", ">3000 Euro", "do not tell",
                                        "13-28 months", "29-52 months",
                                        ">52 months",
                                        "Years of Education",
                                        "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))),
         group = factor(group, levels = c("Contraception", "Age", "Income",
                                          "Relationship Duration","Years of Education",
                                          "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.06))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.06, 0.06), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```

### Sexual Frequency (Penetrative Intercourse) {.tabset}
#### Model
```{r}
m_hc_sexfreqpen_controlled = brm(diary_sex_active_sex_sum ~
                        offset(log(number_of_days)) +
                        contraception_hormonal +
                        age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
                data = data, family = poisson(),
                file = "m_hc_sexfreqpen_controlled")
```

#### Summary
```{r}
kable(tidy(m_hc_sexfreqpen_controlled, intervals = T, prob = 0.90))
summary(m_hc_sexfreqpen_controlled, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_hc_sexfreqpen_controlled, range = c(-0.05, 0.05), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_hc_sexfreqpen_controlled, range = c(-0.05, 0.05), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_hc_sexfreqpen_controlled,
                    effects = "contraception_hormonal",
                    conditions = data.frame(number_of_days = 1))
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_hc_sexfreqpen_controlled %>%
  spread_draws(b_contraception_hormonalyes,
               b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes,
                        b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_relationship_duration_factor",
                        "Relationship Duration",
                        ifelse(condition %contains% "b_net_income",
                               "Income",
                               NA)),
         group = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Contraception", group),
         condition = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Hormonal Contraception", condition),
         condition = ifelse(condition == "b_age", "Age",
                ifelse(condition == "b_net_incomeeuro_500_1000", "500-1000 Euro",
                ifelse(condition == "b_net_incomeeuro_1000_2000", "1000-2000 Euro",
                ifelse(condition == "b_net_incomeeuro_2000_3000", "2000-3000 Euro",
                ifelse(condition == "b_net_incomeeuro_gt_3000", ">3000 Euro",
                ifelse(condition == "b_net_incomedont_tell", "do not tell",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto28months",
                       "13-28 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto52months",
                       "29-52 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_morethan52months",
                       ">52 months",
                ifelse(condition == "b_education_years", "Years of Education",
                ifelse(condition == "b_bfi_extra", "Extraversion",
                ifelse(condition == "b_bfi_neuro", "Neuroticism",
                ifelse(condition == "b_bfi_agree", "Agreeableness",
                ifelse(condition == "b_bfi_consc", "Conscientiousness",
                ifelse(condition == "b_bfi_open", "Openness",
                ifelse(condition == "b_religiosity", "Religiosity",
                       condition)))))))))))))))),
         group = ifelse(is.na(group), condition, group),
         condition = factor(condition, levels = rev(c("Hormonal Contraception", "Age",
                                        "500-1000 Euro", "1000-2000 Euro",
                                        "2000-3000 Euro", ">3000 Euro", "do not tell",
                                        "13-28 months", "29-52 months",
                                        ">52 months",
                                        "Years of Education",
                                        "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))),
         group = factor(group, levels = c("Contraception", "Age", "Income",
                                          "Relationship Duration","Years of Education",
                                          "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.05))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.05, 0.05), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```

### Masturabation Frequency {.tabset}
#### Model
```{r}
m_hc_masfreq_controlled = brm(diary_masturbation_sum ~
                        offset(log(number_of_days)) +
                        contraception_hormonal +
                        age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
                data = data, family = poisson(),
                file = "m_hc_masfreq_controlled")
```

#### Summary
```{r}
kable(tidy(m_hc_masfreq_controlled, intervals = T, prob = 0.90))
summary(m_hc_masfreq_controlled, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_hc_masfreq_controlled, range = c(-0.05, 0.05), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_hc_masfreq_controlled, range = c(-0.05, 0.05), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_hc_masfreq_controlled,
                    effects = "contraception_hormonal",
                    conditions = data.frame(number_of_days = 1))
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_hc_masfreq_controlled %>%
  spread_draws(b_contraception_hormonalyes,
               b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes,
                        b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_relationship_duration_factor",
                        "Relationship Duration",
                        ifelse(condition %contains% "b_net_income",
                               "Income",
                               NA)),
         group = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Contraception", group),
         condition = ifelse(condition %contains% "b_contraception_hormonalyes",
                        "Hormonal Contraception", condition),
         condition = ifelse(condition == "b_age", "Age",
                ifelse(condition == "b_net_incomeeuro_500_1000", "500-1000 Euro",
                ifelse(condition == "b_net_incomeeuro_1000_2000", "1000-2000 Euro",
                ifelse(condition == "b_net_incomeeuro_2000_3000", "2000-3000 Euro",
                ifelse(condition == "b_net_incomeeuro_gt_3000", ">3000 Euro",
                ifelse(condition == "b_net_incomedont_tell", "do not tell",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto28months",
                       "13-28 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto52months",
                       "29-52 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_morethan52months",
                       ">52 months",
                ifelse(condition == "b_education_years", "Years of Education",
                ifelse(condition == "b_bfi_extra", "Extraversion",
                ifelse(condition == "b_bfi_neuro", "Neuroticism",
                ifelse(condition == "b_bfi_agree", "Agreeableness",
                ifelse(condition == "b_bfi_consc", "Conscientiousness",
                ifelse(condition == "b_bfi_open", "Openness",
                ifelse(condition == "b_religiosity", "Religiosity",
                       condition)))))))))))))))),
         group = ifelse(is.na(group), condition, group),
         condition = factor(condition, levels = rev(c("Hormonal Contraception", "Age",
                                        "500-1000 Euro", "1000-2000 Euro",
                                        "2000-3000 Euro", ">3000 Euro", "do not tell",
                                        "13-28 months", "29-52 months",
                                        ">52 months",
                                        "Years of Education",
                                        "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))),
         group = factor(group, levels = c("Contraception", "Age", "Income",
                                          "Relationship Duration","Years of Education",
                                          "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.05))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.05, 0.05), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```

## Effects of (In)Congruenct HC Use {.tabset}
### Attractiveness of Partner  {.tabset}
#### Model
```{r}
m_congruency_atrr = brm(attractiveness_partner ~
                          contraception_hormonal * congruent_contraception,
                data = data, family = gaussian(),
                file = "m_congruency_atrr")
```

#### Summary
```{r}
kable(tidy(m_congruency_atrr, intervals = T, prob = 0.90))
summary(m_congruency_atrr, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_congruency_atrr, range = c(-0.07, 0.07), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_congruency_atrr, range = c(-0.07, 0.07), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_congruency_atrr, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_congruency_atrr %>%
  spread_draws(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "ontraception",
                        "Contraception", NA),
         condition = ifelse(condition == "b_contraception_hormonalyes",
                        "Hormonal Contraception",
                     ifelse(condition == "b_congruent_contraception1",
                            "Congruent Contraception",
                     ifelse(condition == "b_contraception_hormonalyes:congruent_contraception1",
                            "Interaction Hormonal Contracpetion and Congruent Contraception",
                            condition))),
         condition = factor(condition, levels = rev(c("Hormonal Contraception",
                                            "Congruent Contraception",
                                            "Interaction Hormonal Contracpetion and Congruent Contraception")))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.07))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.07, 0.07), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors") +
  xlim (-0.6, 0.6)


  
```

### Relationship Satisfaction {.tabset}
#### Model
```{r}
m_congruency_relsat = brm(relationship_satisfaction ~
                          contraception_hormonal * congruent_contraception,
                data = data, family = gaussian(),
                file = "m_congruency_relsat")
```

#### Summary
```{r}
kable(tidy(m_congruency_relsat, intervals = T, prob = 0.90))
summary(m_congruency_relsat, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_congruency_relsat, range = c(-0.04, 0.04), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_congruency_relsat, range = c(-0.04, 0.04), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_congruency_relsat, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_congruency_relsat %>%
  spread_draws(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "ontraception",
                        "Contraception", NA),
         condition = ifelse(condition == "b_contraception_hormonalyes",
                        "Hormonal Contraception",
                     ifelse(condition == "b_congruent_contraception1",
                            "Congruent Contraception",
                     ifelse(condition == "b_contraception_hormonalyes:congruent_contraception1",
                            "Interaction Hormonal Contracpetion and Congruent Contraception",
                            condition))),
         condition = factor(condition, levels = rev(c("Hormonal Contraception",
                                            "Congruent Contraception",
                                            "Interaction Hormonal Contracpetion and Congruent Contraception")))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.04))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.04, 0.04), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors") +
  xlim (-0.6, 0.6)

```

### Sexual Satisfaction {.tabset}
#### Model
```{r}
m_congruency_sexsat = brm(satisfaction_sexual_intercourse ~
                          contraception_hormonal * congruent_contraception,
                data = data, family = gaussian(),
                file = "m_congruency_sexsat")
```

#### Summary
```{r}
kable(tidy(m_congruency_sexsat, intervals = T, prob = 0.90))
summary(m_congruency_sexsat, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_congruency_sexsat, range = c(-0.11, 0.11), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_congruency_sexsat, range = c(-0.11, 0.11), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_congruency_sexsat, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_congruency_sexsat %>%
  spread_draws(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "ontraception",
                        "Contraception", NA),
         condition = ifelse(condition == "b_contraception_hormonalyes",
                        "Hormonal Contraception",
                     ifelse(condition == "b_congruent_contraception1",
                            "Congruent Contraception",
                     ifelse(condition == "b_contraception_hormonalyes:congruent_contraception1",
                            "Interaction Hormonal Contracpetion and Congruent Contraception",
                            condition))),
         condition = factor(condition, levels = rev(c("Hormonal Contraception",
                                            "Congruent Contraception",
                                            "Interaction Hormonal Contracpetion and Congruent Contraception")))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.11))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.11, 0.11), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors") +
  xlim (-0.6, 0.6)
```

### Libido {.tabset}
#### Model
```{r}
m_congruency_libido = brm(diary_libido_mean ~
                          contraception_hormonal * congruent_contraception,
                data = data, family = gaussian(),
                file = "m_congruency_libido")

```

#### Summary
```{r}
kable(tidy(m_congruency_libido, intervals = T, prob = 0.90))
summary(m_congruency_libido, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_congruency_libido, range = c(-0.06, 0.06), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_congruency_libido, range = c(-0.06, 0.06), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_congruency_libido, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_congruency_libido %>%
  spread_draws(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "ontraception",
                        "Contraception", NA),
         condition = ifelse(condition == "b_contraception_hormonalyes",
                        "Hormonal Contraception",
                     ifelse(condition == "b_congruent_contraception1",
                            "Congruent Contraception",
                     ifelse(condition == "b_contraception_hormonalyes:congruent_contraception1",
                            "Interaction Hormonal Contracpetion and Congruent Contraception",
                            condition))),
         condition = factor(condition, levels = rev(c("Hormonal Contraception",
                                            "Congruent Contraception",
                                            "Interaction Hormonal Contracpetion and Congruent Contraception")))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.06))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.06, 0.06), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors") +
  xlim (-0.6, 0.6)
```

### Sexual Frequency (Penetrative Intercourse) {.tabset}
#### Model
```{r}
m_congruency_sexfreqpen = brm(diary_sex_active_sex_sum ~ offset(log(number_of_days)) +
                     contraception_hormonal * congruent_contraception,
                data = data, family = poisson(),
                file = "m_congruency_sexfreqpen")
```

#### Summary
```{r}
kable(tidy(m_congruency_sexfreqpen, intervals = T, prob = 0.90))
summary(m_congruency_sexfreqpen, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_congruency_sexfreqpen, range = c(-0.05, 0.05), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_congruency_sexfreqpen, range = c(-0.05, 0.05), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_congruency_sexfreqpen,
                    effects = "contraception_hormonal:congruent_contraception",
                    conditions = data.frame(number_of_days = 1))
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_congruency_sexfreqpen %>%
  spread_draws(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "ontraception",
                        "Contraception", NA),
         condition = ifelse(condition == "b_contraception_hormonalyes",
                        "Hormonal Contraception",
                     ifelse(condition == "b_congruent_contraception1",
                            "Congruent Contraception",
                     ifelse(condition == "b_contraception_hormonalyes:congruent_contraception1",
                            "Interaction Hormonal Contracpetion and Congruent Contraception",
                            condition))),
         condition = factor(condition, levels = rev(c("Hormonal Contraception",
                                            "Congruent Contraception",
                                            "Interaction Hormonal Contracpetion and Congruent Contraception")))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.05))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.05, 0.051), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```

### Masturbation Frequency {.tabset}
#### Model
```{r}
m_congruency_masfreq = brm(diary_masturbation_sum ~ offset(log(number_of_days)) +
                             contraception_hormonal * congruent_contraception,
                data = data, family = poisson(),
                file = "m_congruency_masfreq")
```

#### Summary
```{r}
kable(tidy(m_congruency_masfreq, intervals = T, prob = 0.90))
summary(m_congruency_masfreq, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_congruency_masfreq, range = c(-0.05, 0.05), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_congruency_masfreq, range = c(-0.05, 0.05), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_congruency_masfreq,
                    effects = "contraception_hormonal:congruent_contraception",
                    conditions = data.frame(number_of_days = 1))
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_congruency_masfreq %>%
spread_draws(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "ontraception",
                        "Contraception", NA),
         condition = ifelse(condition == "b_contraception_hormonalyes",
                        "Hormonal Contraception",
                     ifelse(condition == "b_congruent_contraception1",
                            "Congruent Contraception",
                     ifelse(condition == "b_contraception_hormonalyes:congruent_contraception1",
                            "Interaction Hormonal Contracpetion and Congruent Contraception",
                            condition))),
         condition = factor(condition, levels = rev(c("Hormonal Contraception",
                                            "Congruent Contraception",
                                            "Interaction Hormonal Contracpetion and Congruent Contraception")))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.05))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.05, 0.05), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```



## Controlled Models: Effects of (In)Congruenct HC Use {.tabset}
### Attractiveness of Partner  {.tabset}
#### Model
```{r}
m_congruency_atrr_controlled = brm(attractiveness_partner ~
                          contraception_hormonal * congruent_contraception +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
                data = data, family = gaussian(),
                file = "m_congruency_atrr_controlled")
```

#### Summary
```{r}
kable(tidy(m_congruency_atrr_controlled, intervals = T, prob = 0.90))
summary(m_congruency_atrr_controlled, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_congruency_atrr_controlled, range = c(-0.07, 0.07), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_congruency_atrr_controlled, range = c(-0.07, 0.07), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_congruency_atrr_controlled, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_congruency_atrr_controlled %>%
  spread_draws(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`,
               b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`,
                        b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_relationship_duration_factor",
                        "Relationship Duration",
                        ifelse(condition %contains% "b_net_income",
                               "Income",
                               NA)),
         group = ifelse(condition %contains% "ontraception",
                        "Contraception", group),
         condition = ifelse(condition == "b_contraception_hormonalyes",
                        "Hormonal Contraception",
                     ifelse(condition == "b_congruent_contraception1",
                            "Congruent Contraception",
                     ifelse(condition == "b_contraception_hormonalyes:congruent_contraception1",
                            "Interaction Hormonal Contracpetion and Congruent Contraception",
                            condition))),
         condition = ifelse(condition == "b_age", "Age",
                ifelse(condition == "b_net_incomeeuro_500_1000", "500-1000 Euro",
                ifelse(condition == "b_net_incomeeuro_1000_2000", "1000-2000 Euro",
                ifelse(condition == "b_net_incomeeuro_2000_3000", "2000-3000 Euro",
                ifelse(condition == "b_net_incomeeuro_gt_3000", ">3000 Euro",
                ifelse(condition == "b_net_incomedont_tell", "do not tell",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto28months",
                       "13-28 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto52months",
                       "29-52 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_morethan52months",
                       ">52 months",
                ifelse(condition == "b_education_years", "Years of Education",
                ifelse(condition == "b_bfi_extra", "Extraversion",
                ifelse(condition == "b_bfi_neuro", "Neuroticism",
                ifelse(condition == "b_bfi_agree", "Agreeableness",
                ifelse(condition == "b_bfi_consc", "Conscientiousness",
                ifelse(condition == "b_bfi_open", "Openness",
                ifelse(condition == "b_religiosity", "Religiosity",
                       condition)))))))))))))))),
         group = ifelse(is.na(group), condition, group),
         condition = factor(condition, levels = rev(c("Hormonal Contraception",
                                        "Congruent Contraception",
                                        "Interaction Hormonal Contracpetion and Congruent Contraception",
                                        "Age",
                                        "500-1000 Euro", "1000-2000 Euro",
                                        "2000-3000 Euro", ">3000 Euro", "do not tell",
                                        "13-28 months", "29-52 months",
                                        ">52 months",
                                        "Years of Education",
                                        "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))),
         group = factor(group, levels = c("Contraception", "Age", "Income",
                                          "Relationship Duration","Years of Education",
                                          "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.07))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.07, 0.07), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```

### Relationship Satisfaction {.tabset}
#### Model
```{r}
m_congruency_relsat_controlled = brm(relationship_satisfaction ~
                          contraception_hormonal * congruent_contraception +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
                data = data, family = gaussian(),
                file = "m_congruency_relsat_controlled")
```

#### Summary
```{r}
kable(tidy(m_congruency_relsat_controlled, intervals = T, prob = 0.90))
summary(m_congruency_relsat_controlled, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_congruency_relsat_controlled, range = c(-0.04, 0.04), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_congruency_relsat_controlled, range = c(-0.04, 0.04), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_congruency_relsat_controlled, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_congruency_relsat_controlled %>%
  
  spread_draws(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`,
               b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`,
                        b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_relationship_duration_factor",
                        "Relationship Duration",
                        ifelse(condition %contains% "b_net_income",
                               "Income",
                               NA)),
         group = ifelse(condition %contains% "ontraception",
                        "Contraception", group),
         condition = ifelse(condition == "b_contraception_hormonalyes",
                        "Hormonal Contraception",
                     ifelse(condition == "b_congruent_contraception1",
                            "Congruent Contraception",
                     ifelse(condition == "b_contraception_hormonalyes:congruent_contraception1",
                            "Interaction Hormonal Contracpetion and Congruent Contraception",
                            condition))),
         condition = ifelse(condition == "b_age", "Age",
                ifelse(condition == "b_net_incomeeuro_500_1000", "500-1000 Euro",
                ifelse(condition == "b_net_incomeeuro_1000_2000", "1000-2000 Euro",
                ifelse(condition == "b_net_incomeeuro_2000_3000", "2000-3000 Euro",
                ifelse(condition == "b_net_incomeeuro_gt_3000", ">3000 Euro",
                ifelse(condition == "b_net_incomedont_tell", "do not tell",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto28months",
                       "13-28 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto52months",
                       "29-52 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_morethan52months",
                       ">52 months",
                ifelse(condition == "b_education_years", "Years of Education",
                ifelse(condition == "b_bfi_extra", "Extraversion",
                ifelse(condition == "b_bfi_neuro", "Neuroticism",
                ifelse(condition == "b_bfi_agree", "Agreeableness",
                ifelse(condition == "b_bfi_consc", "Conscientiousness",
                ifelse(condition == "b_bfi_open", "Openness",
                ifelse(condition == "b_religiosity", "Religiosity",
                       condition)))))))))))))))),
         group = ifelse(is.na(group), condition, group),
         condition = factor(condition, levels = rev(c("Hormonal Contraception",
                                        "Congruent Contraception",
                                        "Interaction Hormonal Contracpetion and Congruent Contraception",
                                        "Age",
                                        "500-1000 Euro", "1000-2000 Euro",
                                        "2000-3000 Euro", ">3000 Euro", "do not tell",
                                        "13-28 months", "29-52 months",
                                        ">52 months",
                                        "Years of Education",
                                        "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))),
         group = factor(group, levels = c("Contraception", "Age", "Income",
                                          "Relationship Duration","Years of Education",
                                          "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.04))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.04, 0.04), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```

### Sexual Satisfaction {.tabset}
#### Model
```{r}
m_congruency_sexsat_controlled = brm(satisfaction_sexual_intercourse ~
                          contraception_hormonal * congruent_contraception +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
                data = data, family = gaussian(),
                file = "m_congruency_sexsat_controlled")
```

#### Summary
```{r}
kable(tidy(m_congruency_sexsat_controlled, intervals = T, prob = 0.90))
summary(m_congruency_sexsat_controlled, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_congruency_sexsat_controlled, range = c(-0.11, 0.11), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_congruency_sexsat_controlled, range = c(-0.11, 0.11), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_congruency_sexsat_controlled, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_congruency_sexsat_controlled %>%
  
  spread_draws(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`,
               b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`,
                        b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_relationship_duration_factor",
                        "Relationship Duration",
                        ifelse(condition %contains% "b_net_income",
                               "Income",
                               NA)),
         group = ifelse(condition %contains% "ontraception",
                        "Contraception", group),
         condition = ifelse(condition == "b_contraception_hormonalyes",
                        "Hormonal Contraception",
                     ifelse(condition == "b_congruent_contraception1",
                            "Congruent Contraception",
                     ifelse(condition == "b_contraception_hormonalyes:congruent_contraception1",
                            "Interaction Hormonal Contracpetion and Congruent Contraception",
                            condition))),
         condition = ifelse(condition == "b_age", "Age",
                ifelse(condition == "b_net_incomeeuro_500_1000", "500-1000 Euro",
                ifelse(condition == "b_net_incomeeuro_1000_2000", "1000-2000 Euro",
                ifelse(condition == "b_net_incomeeuro_2000_3000", "2000-3000 Euro",
                ifelse(condition == "b_net_incomeeuro_gt_3000", ">3000 Euro",
                ifelse(condition == "b_net_incomedont_tell", "do not tell",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto28months",
                       "13-28 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto52months",
                       "29-52 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_morethan52months",
                       ">52 months",
                ifelse(condition == "b_education_years", "Years of Education",
                ifelse(condition == "b_bfi_extra", "Extraversion",
                ifelse(condition == "b_bfi_neuro", "Neuroticism",
                ifelse(condition == "b_bfi_agree", "Agreeableness",
                ifelse(condition == "b_bfi_consc", "Conscientiousness",
                ifelse(condition == "b_bfi_open", "Openness",
                ifelse(condition == "b_religiosity", "Religiosity",
                       condition)))))))))))))))),
         group = ifelse(is.na(group), condition, group),
         condition = factor(condition, levels = rev(c("Hormonal Contraception",
                                        "Congruent Contraception",
                                        "Interaction Hormonal Contracpetion and Congruent Contraception",
                                        "Age",
                                        "500-1000 Euro", "1000-2000 Euro",
                                        "2000-3000 Euro", ">3000 Euro", "do not tell",
                                        "13-28 months", "29-52 months",
                                        ">52 months",
                                        "Years of Education",
                                        "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))),
         group = factor(group, levels = c("Contraception", "Age", "Income",
                                          "Relationship Duration","Years of Education",
                                          "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.11))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.11, 0.11), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```

### Libido {.tabset}
#### Model
```{r}
m_congruency_libido_controlled = brm(diary_libido_mean ~
                          contraception_hormonal * congruent_contraception +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
                data = data, family = gaussian(),
                file = "m_congruency_libido_controlled")

```

#### Summary
```{r}
kable(tidy(m_congruency_libido_controlled, intervals = T, prob = 0.90))
summary(m_congruency_libido_controlled, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_congruency_libido_controlled, range = c(-0.06, 0.06), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_congruency_libido_controlled, range = c(-0.06, 0.06), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_congruency_libido_controlled, ask = FALSE)
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_congruency_libido_controlled %>%
  
  spread_draws(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`,
               b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`,
                        b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_relationship_duration_factor",
                        "Relationship Duration",
                        ifelse(condition %contains% "b_net_income",
                               "Income",
                               NA)),
         group = ifelse(condition %contains% "ontraception",
                        "Contraception", group),
         condition = ifelse(condition == "b_contraception_hormonalyes",
                        "Hormonal Contraception",
                     ifelse(condition == "b_congruent_contraception1",
                            "Congruent Contraception",
                     ifelse(condition == "b_contraception_hormonalyes:congruent_contraception1",
                            "Interaction Hormonal Contracpetion and Congruent Contraception",
                            condition))),
         condition = ifelse(condition == "b_age", "Age",
                ifelse(condition == "b_net_incomeeuro_500_1000", "500-1000 Euro",
                ifelse(condition == "b_net_incomeeuro_1000_2000", "1000-2000 Euro",
                ifelse(condition == "b_net_incomeeuro_2000_3000", "2000-3000 Euro",
                ifelse(condition == "b_net_incomeeuro_gt_3000", ">3000 Euro",
                ifelse(condition == "b_net_incomedont_tell", "do not tell",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto28months",
                       "13-28 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto52months",
                       "29-52 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_morethan52months",
                       ">52 months",
                ifelse(condition == "b_education_years", "Years of Education",
                ifelse(condition == "b_bfi_extra", "Extraversion",
                ifelse(condition == "b_bfi_neuro", "Neuroticism",
                ifelse(condition == "b_bfi_agree", "Agreeableness",
                ifelse(condition == "b_bfi_consc", "Conscientiousness",
                ifelse(condition == "b_bfi_open", "Openness",
                ifelse(condition == "b_religiosity", "Religiosity",
                       condition)))))))))))))))),
         group = ifelse(is.na(group), condition, group),
         condition = factor(condition, levels = rev(c("Hormonal Contraception",
                                        "Congruent Contraception",
                                        "Interaction Hormonal Contracpetion and Congruent Contraception",
                                        "Age",
                                        "500-1000 Euro", "1000-2000 Euro",
                                        "2000-3000 Euro", ">3000 Euro", "do not tell",
                                        "13-28 months", "29-52 months",
                                        ">52 months",
                                        "Years of Education",
                                        "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))),
         group = factor(group, levels = c("Contraception", "Age", "Income",
                                          "Relationship Duration","Years of Education",
                                          "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.06))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.06, 0.06), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```

### Sexual Frequency (Penetrative Intercourse) {.tabset}
#### Model
```{r}
m_congruency_sexfreqpen_controlled = brm(diary_sex_active_sex_sum ~
                        offset(log(number_of_days)) +
                        contraception_hormonal * congruent_contraception +
                        age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
                data = data, family = poisson(),
                file = "m_congruency_sexfreqpen_controlled")
```

#### Summary
```{r}
kable(tidy(m_congruency_sexfreqpen_controlled, intervals = T, prob = 0.90))
summary(m_congruency_sexfreqpen_controlled, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_congruency_sexfreqpen_controlled, range = c(-0.05, 0.05),
                      ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_congruency_sexfreqpen_controlled, range = c(-0.05, 0.05), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_congruency_sexfreqpen_controlled,
                    effects = "contraception_hormonal:congruent_contraception",
                    conditions = data.frame(number_of_days = 1))
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_congruency_sexfreqpen_controlled %>%

  spread_draws(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`,
               b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`,
                        b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_relationship_duration_factor",
                        "Relationship Duration",
                        ifelse(condition %contains% "b_net_income",
                               "Income",
                               NA)),
         group = ifelse(condition %contains% "ontraception",
                        "Contraception", group),
         condition = ifelse(condition == "b_contraception_hormonalyes",
                        "Hormonal Contraception",
                     ifelse(condition == "b_congruent_contraception1",
                            "Congruent Contraception",
                     ifelse(condition == "b_contraception_hormonalyes:congruent_contraception1",
                            "Interaction Hormonal Contracpetion and Congruent Contraception",
                            condition))),
         condition = ifelse(condition == "b_age", "Age",
                ifelse(condition == "b_net_incomeeuro_500_1000", "500-1000 Euro",
                ifelse(condition == "b_net_incomeeuro_1000_2000", "1000-2000 Euro",
                ifelse(condition == "b_net_incomeeuro_2000_3000", "2000-3000 Euro",
                ifelse(condition == "b_net_incomeeuro_gt_3000", ">3000 Euro",
                ifelse(condition == "b_net_incomedont_tell", "do not tell",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto28months",
                       "13-28 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto52months",
                       "29-52 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_morethan52months",
                       ">52 months",
                ifelse(condition == "b_education_years", "Years of Education",
                ifelse(condition == "b_bfi_extra", "Extraversion",
                ifelse(condition == "b_bfi_neuro", "Neuroticism",
                ifelse(condition == "b_bfi_agree", "Agreeableness",
                ifelse(condition == "b_bfi_consc", "Conscientiousness",
                ifelse(condition == "b_bfi_open", "Openness",
                ifelse(condition == "b_religiosity", "Religiosity",
                       condition)))))))))))))))),
         group = ifelse(is.na(group), condition, group),
         condition = factor(condition, levels = rev(c("Hormonal Contraception",
                                        "Congruent Contraception",
                                        "Interaction Hormonal Contracpetion and Congruent Contraception",
                                        "Age",
                                        "500-1000 Euro", "1000-2000 Euro",
                                        "2000-3000 Euro", ">3000 Euro", "do not tell",
                                        "13-28 months", "29-52 months",
                                        ">52 months",
                                        "Years of Education",
                                        "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))),
         group = factor(group, levels = c("Contraception", "Age", "Income",
                                          "Relationship Duration","Years of Education",
                                          "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.05))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.05, 0.05), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```

### Masturbation Frequency {.tabset}
#### Model
```{r}
m_congruency_masfreq_controlled = brm(diary_masturbation_sum ~
                        offset(log(number_of_days)) +
                        contraception_hormonal * congruent_contraception +
                        age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
                data = data, family = poisson(),
                file = "m_congruency_masfreq_controlled")
```

#### Summary
```{r}
kable(tidy(m_congruency_masfreq_controlled, intervals = T, prob = 0.90))
summary(m_congruency_masfreq_controlled, intervals = T, prob = 0.90)
```

#### Comparison with ROPE
```{r}
plot(equivalence_test(m_congruency_masfreq_controlled, range = c(-0.05, 0.05), ci = 0.90,
                      parameters = "contraception"))
equivalence_test(m_congruency_masfreq_controlled, range = c(-0.05, 0.05), ci = 0.90,
                      parameters = "contraception")
```

#### Plots
```{r}
conditional_effects(m_congruency_masfreq_controlled,
                    effects = "contraception_hormonal:congruent_contraception",
                    conditions = data.frame(number_of_days = 1))
```

#### Forest Plot for Effect Sizes {.active}
```{r}
m_congruency_masfreq_controlled %>%

  spread_draws(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`,
               b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity) %>%
  pivot_longer(cols = c(b_contraception_hormonalyes, b_congruent_contraception1,
               `b_contraception_hormonalyes:congruent_contraception1`,
                        b_age,
               b_net_incomeeuro_500_1000, b_net_incomeeuro_1000_2000, 
               b_net_incomeeuro_2000_3000, b_net_incomeeuro_gt_3000, b_net_incomedont_tell,
               b_relationship_duration_factorPartnered_upto28months,
               b_relationship_duration_factorPartnered_upto52months,
               b_relationship_duration_factorPartnered_morethan52months,
               b_education_years,
               b_bfi_extra, b_bfi_neuro, b_bfi_agree, b_bfi_consc, b_bfi_open, 
               b_religiosity),
               names_to = "condition",
               values_to = "r_condition") %>%
  mutate(condition_mean = r_condition,
         group = ifelse(condition %contains% "b_relationship_duration_factor",
                        "Relationship Duration",
                        ifelse(condition %contains% "b_net_income",
                               "Income",
                               NA)),
         group = ifelse(condition %contains% "ontraception",
                        "Contraception", group),
         condition = ifelse(condition == "b_contraception_hormonalyes",
                        "Hormonal Contraception",
                     ifelse(condition == "b_congruent_contraception1",
                            "Congruent Contraception",
                     ifelse(condition == "b_contraception_hormonalyes:congruent_contraception1",
                            "Interaction Hormonal Contracpetion and Congruent Contraception",
                            condition))),
         condition = ifelse(condition == "b_age", "Age",
                ifelse(condition == "b_net_incomeeuro_500_1000", "500-1000 Euro",
                ifelse(condition == "b_net_incomeeuro_1000_2000", "1000-2000 Euro",
                ifelse(condition == "b_net_incomeeuro_2000_3000", "2000-3000 Euro",
                ifelse(condition == "b_net_incomeeuro_gt_3000", ">3000 Euro",
                ifelse(condition == "b_net_incomedont_tell", "do not tell",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto28months",
                       "13-28 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_upto52months",
                       "29-52 months",
                ifelse(condition == "b_relationship_duration_factorPartnered_morethan52months",
                       ">52 months",
                ifelse(condition == "b_education_years", "Years of Education",
                ifelse(condition == "b_bfi_extra", "Extraversion",
                ifelse(condition == "b_bfi_neuro", "Neuroticism",
                ifelse(condition == "b_bfi_agree", "Agreeableness",
                ifelse(condition == "b_bfi_consc", "Conscientiousness",
                ifelse(condition == "b_bfi_open", "Openness",
                ifelse(condition == "b_religiosity", "Religiosity",
                       condition)))))))))))))))),
         group = ifelse(is.na(group), condition, group),
         condition = factor(condition, levels = rev(c("Hormonal Contraception",
                                        "Congruent Contraception",
                                        "Interaction Hormonal Contracpetion and Congruent Contraception",
                                        "Age",
                                        "500-1000 Euro", "1000-2000 Euro",
                                        "2000-3000 Euro", ">3000 Euro", "do not tell",
                                        "13-28 months", "29-52 months",
                                        ">52 months",
                                        "Years of Education",
                                        "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))),
         group = factor(group, levels = c("Contraception", "Age", "Income",
                                          "Relationship Duration","Years of Education",
                                          "Extraversion", "Neuroticism", "Agreeableness",
                                         "Conscientiousness","Openness","Religiosity"))) %>%
  ggplot(aes(y = condition,
             x = condition_mean,
             fill = stat(abs(x) < 0.05))) +
  stat_halfeye() +
  geom_vline(xintercept = c(-0.05, 0.05), linetype = "dotted") +
  apatheme +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("gray80", "skyblue")) +
  labs(x = "Effect Size Estimates", y = "Predictors")
```

### Model Comparison
#### Attractiveness of partner
```{r}
compare_models = loo(m_hc_atrr, m_congruency_atrr)
```
LooIC first model: `r round(compare_models$loos$m_hc_atrr$estimates[3,1], 2)`

LooIC second model: `r round(compare_models$loos$m_congruency_atrr$estimates[3,1], 2)`

Model Comparisons: The difference between models is `r round(compare_models$diffs[2,1], 2)` compared to a standard error of `r round(compare_models$diffs[2,2], 2)`
```{r}
compare_models
```

#### Relationship satisfaction
```{r}
compare_models = loo(m_hc_relsat, m_congruency_relsat)
```
LooIC first model: `r round(compare_models$loos$m_hc_relsat$estimates[3,1], 2)`

LooIC second model: `r round(compare_models$loos$m_congruency_relsat$estimates[3,1], 2)`

Model Comparisons: The difference between models is `r round(compare_models$diffs[2,1], 2)` compared to a standard error of `r round(compare_models$diffs[2,2], 2)`
```{r}
compare_models
```

#### Sexual Satisfaction
```{r}
m_hc_sexsat$data$satisfaction_sexual_intercourse =
  as.numeric(m_hc_sexsat$data$satisfaction_sexual_intercourse)
m_congruency_sexsat$data$satisfaction_sexual_intercourse =
  as.numeric(m_congruency_sexsat$data$satisfaction_sexual_intercourse)

compare_models = loo(m_hc_sexsat, m_congruency_sexsat)
```
LooIC first model: `r round(compare_models$loos$m_hc_sexsat$estimates[3,1], 2)`

LooIC second model: `r round(compare_models$loos$m_congruency_sexsat$estimates[3,1], 2)`

Model Comparisons: The difference between models is `r round(compare_models$diffs[2,1], 2)` compared to a standard error of `r round(compare_models$diffs[2,2], 2)`
```{r}
compare_models
```
