==> 7_analyses_sensitivity.Rmd <==
---
title: "Sensitivity Analysis"
output:
  html_document:
    toc: true
    toc_depth: 4
    toc_float: true
    code_folding: 'hide'
    self_contained: false
---
  
  
## Data 
```{r results='hide',message=F,warning=F}
source("0_helpers.R")

load("data/cleaned_selected_wrangled.rdata")
```

## Preparations {.tabset}
### Change factors to numerics, handcode interaction
```{r}
data = data %>%
  mutate(contraception_hormonal_numeric = ifelse(contraception_hormonal == "yes",
                                                 1,
                                                 ifelse(contraception_hormonal == "no",
                                                        0, NA)),
         congruent_contraception_numeric = ifelse(congruent_contraception == "0",
                                                  0,
                                                  ifelse(congruent_contraception == "1",
                                                         1, NA)),
         hc_con_interaction = ifelse(is.na(congruent_contraception), NA,
                                     ifelse(contraception_hormonal == "yes" &
                                              congruent_contraception == "1", 1, 0)))
```

### Covariates
```{r}
covariates = list("age",
                   "net_incomeeuro_500_1000", "net_incomeeuro_1000_2000",
                   "net_incomeeuro_2000_3000", "net_incomeeuro_gt_3000",
                    "net_incomedont_tell",
                   "relationship_duration_factorPartnered_upto28months",
                   "relationship_duration_factorPartnered_upto52months",
                   "relationship_duration_factorPartnered_morethan52months",
                   "education_years", "bfi_extra", "bfi_neuro", "bfi_agree",
                   "bfi_consc", "bfi_open", "religiosity")
names(covariates) = c("age",
                   "net_incomeeuro_500_1000", "net_incomeeuro_1000_2000",
                   "net_incomeeuro_2000_3000", "net_incomeeuro_gt_3000",
                    "net_incomedont_tell",
                   "relationship_duration_factorPartnered_upto28months",
                   "relationship_duration_factorPartnered_upto52months",
                   "relationship_duration_factorPartnered_morethan52months",
                   "education_years", "bfi_extra", "bfi_neuro", "bfi_agree",
                   "bfi_consc", "bfi_open", "religiosity")
```

## Effects of Hormonal Contraception  {.tabset}
### Attractiveness of Partner {.tabset}
#### Uncontrolled Model {.tabset}
##### Model
```{r}
m_hc_atrr = lm(attractiveness_partner ~ contraception_hormonal_numeric,
               data = data)
summary(m_hc_atrr)
tidy(m_hc_atrr)
```

##### Sensitivity Analyses
```{r}
m_hc_atrr_sensitivity <- sensemakr(model = m_hc_atrr, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_atrr_sensitivity
summary(m_hc_atrr_sensitivity)
```

#### Controlled Model {.tabset}
##### Model
```{r}
m_hc_atrr = lm(attractiveness_partner ~ contraception_hormonal_numeric +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
               data = data)
summary(m_hc_atrr)
tidy(m_hc_atrr)
```

##### Sensitivity Analyses
```{r}
m_hc_atrr_sensitivity <- sensemakr(model = m_hc_atrr, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_atrr_sensitivity
summary(m_hc_atrr_sensitivity)
```

### Relationship Satisfaction {.tabset}
#### Uncontrolled Model {.tabset}
##### Model
```{r}
m_hc_relsat = lm(relationship_satisfaction ~ contraception_hormonal_numeric,
               data = data)
summary(m_hc_relsat)
tidy(m_hc_relsat)
```

##### Sensitivity Analyses
```{r}
m_hc_relsat_sensitivity <- sensemakr(model = m_hc_relsat, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_relsat_sensitivity
summary(m_hc_relsat_sensitivity)
```

#### Controlled Model {.tabset}
##### Model
```{r}
m_hc_relsat = lm(relationship_satisfaction ~ contraception_hormonal_numeric +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
               data = data)
summary(m_hc_relsat)
tidy(m_hc_relsat)
```

##### Sensitivity Analyses
```{r}
m_hc_relsat_sensitivity <- sensemakr(model = m_hc_relsat, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_relsat_sensitivity
summary(m_hc_relsat_sensitivity)
```

### Sexual Satisfaction {.tabset}
#### Uncontrolled Model {.tabset}
##### Model
```{r}
m_hc_sexsat = lm(satisfaction_sexual_intercourse ~ contraception_hormonal_numeric +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
               data = data)
summary(m_hc_sexsat)
tidy(m_hc_sexsat)
```

##### Sensitivity Analyses
```{r}
m_hc_sexsat_sensitivity <- sensemakr(model = m_hc_sexsat, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_sexsat_sensitivity
summary(m_hc_sexsat_sensitivity)
```

#### Controlled Model {.tabset}
##### Model
```{r}
m_hc_sexsat = lm(satisfaction_sexual_intercourse ~ contraception_hormonal_numeric +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
               data = data)
summary(m_hc_sexsat)
tidy(m_hc_sexsat)
```

##### Sensitivity Analyses
```{r}
m_hc_sexsat_sensitivity <- sensemakr(model = m_hc_sexsat, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_sexsat_sensitivity
summary(m_hc_sexsat_sensitivity)
```

### Libido {.tabset}
#### Uncontrolled Model {.tabset}
##### Model
```{r}
m_hc_libido = lm(diary_libido_mean ~ contraception_hormonal_numeric,
               data = data)
qplot(residuals(m_hc_libido))
summary(m_hc_libido)
tidy(m_hc_libido)
```

##### Sensitivity Analyses
```{r}
m_hc_libido_sensitivity <- sensemakr(model = m_hc_libido, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_libido_sensitivity
summary(m_hc_libido_sensitivity)
```

#### Controlled Model {.tabset}
##### Model
```{r}
m_hc_libido = lm(diary_libido_mean ~ contraception_hormonal_numeric +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
               data = data)
summary(m_hc_libido)
tidy(m_hc_libido)
```

##### Sensitivity Analyses
```{r}
m_hc_libido_sensitivity <- sensemakr(model = m_hc_libido, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_libido_sensitivity
summary(m_hc_libido_sensitivity)
```


### Sexual Frequency{.tabset}
#### Uncontrolled Model {.tabset}
##### Model
```{r}
m_hc_sexfreqpen = lm(diary_sex_active_sex_mean ~ contraception_hormonal_numeric,
               data = data)
qplot(residuals(m_hc_sexfreqpen))
summary(m_hc_sexfreqpen)
tidy(m_hc_sexfreqpen, conf.int = T)
```

##### Sensitivity Analyses
```{r}
m_hc_sexfreqpen_sensitivity <- sensemakr(model = m_hc_sexfreqpen,
                                         treatment = "contraception_hormonal_numeric",
                                        kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05,
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_sexfreqpen_sensitivity
summary(m_hc_sexfreqpen_sensitivity)
```



#### Controlled Model {.tabset}
##### Model
```{r}
m_hc_sexfreqpen = lm(diary_sex_active_sex_mean ~ contraception_hormonal_numeric +
                       age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
               data = data)
qplot(residuals(m_hc_sexfreqpen))
summary(m_hc_sexfreqpen)
tidy(m_hc_sexfreqpen, conf.int = T)
```

##### Sensitivity Analyses
```{r}
m_hc_sexfreqpen_sensitivity <- sensemakr(model = m_hc_sexfreqpen,
                                         treatment = "contraception_hormonal_numeric",
                                         benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05,
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_sexfreqpen_sensitivity
summary(m_hc_sexfreqpen_sensitivity)
```


### Masturbation Frequency{.tabset}
#### Uncontrolled Model {.tabset}
##### Model
```{r}
m_hc_masfreqpen = lm(diary_masturbation_mean ~ contraception_hormonal_numeric,
                     data = data)
qplot(residuals(m_hc_masfreqpen))
summary(m_hc_masfreqpen)
tidy(m_hc_masfreqpen, conf.int = T)
```

##### Sensitivity Analyses
```{r}
m_hc_masfreqpen_sensitivity <- sensemakr(model = m_hc_masfreqpen,
                                         treatment = "contraception_hormonal_numeric",
                                         kd = 1:3, #these arguments parameterize how many times
                                         #stronger the confounder is related to the
                                         #treatment
                                         ky = 1:3, #these arguments parameterize how many times
                                         #stronger the confounder is related to the outcome
                                         q = 1, #fraction of the effect estimate that would have to be
                                         #explained away to be problematic. Setting q = 1,
                                         #means that a reduction of 100% of the current effect
                                         #estimate, that is, a true effect of zero, would be
                                         #deemed problematic.
                                         alpha = 0.05,
                                         reduce = TRUE #confounder reduce absolute effect size
)


m_hc_masfreqpen_sensitivity
summary(m_hc_masfreqpen_sensitivity)
```



#### Controlled Model {.tabset}
##### Model
```{r}
m_hc_masfreqpen = lm(diary_masturbation_mean ~ contraception_hormonal_numeric +
                       age + net_income + relationship_duration_factor +
                       education_years +
                       bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                       religiosity,
                     data = data)
qplot(residuals(m_hc_masfreqpen))
summary(m_hc_masfreqpen)
tidy(m_hc_masfreqpen, conf.int = T)
```

##### Sensitivity Analyses
```{r}
m_hc_masfreqpen_sensitivity <- sensemakr(model = m_hc_masfreqpen,
                                         treatment = "contraception_hormonal_numeric",
                                         benchmark_covariates = covariates, #covariates that will be
                                         #used to bound the
                                         #plausible strength of the
                                         #unobserved confounders
                                         kd = 1:3, #these arguments parameterize how many times
                                         #stronger the confounder is related to the
                                         #treatment
                                         ky = 1:3, #these arguments parameterize how many times
                                         #stronger the confounder is related to the outcome
                                         q = 1, #fraction of the effect estimate that would have to be
                                         #explained away to be problematic. Setting q = 1,
                                         #means that a reduction of 100% of the current effect
                                         #estimate, that is, a true effect of zero, would be
                                         #deemed problematic.
                                         alpha = 0.05,
                                         reduce = TRUE #confounder reduce absolute effect size
)


m_hc_masfreqpen_sensitivity
summary(m_hc_masfreqpen_sensitivity)
```


## HC, Congruent use of HC, and their interaction {.tabset}
### Attractiveness of Partner {.tabset}
#### Uncontrolled Model {.tabset}
##### Model
```{r}
m_con_atrr = lm(attractiveness_partner ~ contraception_hormonal_numeric +
                 congruent_contraception_numeric + hc_con_interaction,
               data = data)
summary(m_con_atrr)
tidy(m_con_atrr)
```

##### Sensitivity Analyses {.tabset}
###### HC
```{r}
m_con_atrr_sensitivity_hc <- sensemakr(model = m_con_atrr, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_atrr_sensitivity_hc
summary(m_con_atrr_sensitivity_hc)
```


###### Congruency
```{r}
m_con_atrr_sensitivity_con <- sensemakr(model = m_con_atrr, #model
                                treatment = "congruent_contraception_numeric", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_atrr_sensitivity_con
summary(m_con_atrr_sensitivity_con)
```

###### Interaction
```{r}
m_con_atrr_sensitivity_interaction <- sensemakr(model = m_con_atrr, #model
                                treatment = "hc_con_interaction", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_atrr_sensitivity_interaction
summary(m_con_atrr_sensitivity_interaction)
```

#### Controlled Model {.tabset}
##### Model
```{r}
m_con_atrr = lm(attractiveness_partner ~ contraception_hormonal_numeric +
                 congruent_contraception_numeric + hc_con_interaction +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
               data = data)
summary(m_con_atrr)
tidy(m_con_atrr)
```

##### Sensitivity Analyses {.tabset}
###### HC
```{r}
m_con_atrr_sensitivity_hc <- sensemakr(model = m_con_atrr, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_atrr_sensitivity_hc
summary(m_con_atrr_sensitivity_hc)
```


###### Congruency
```{r}
m_con_atrr_sensitivity_con <- sensemakr(model = m_con_atrr, #model
                                treatment = "congruent_contraception_numeric", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_atrr_sensitivity_con
summary(m_con_atrr_sensitivity_con)
```

###### Interaction
```{r}
m_con_atrr_sensitivity_interaction <- sensemakr(model = m_con_atrr, #model
                                treatment = "hc_con_interaction", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_atrr_sensitivity_interaction
summary(m_con_atrr_sensitivity_interaction)
```

### Relationship Satisfaction {.tabset}
#### Uncntrolled Model {.tabset}
##### Model
```{r}
m_con_relsat = lm(relationship_satisfaction ~ contraception_hormonal_numeric +
                 congruent_contraception_numeric + hc_con_interaction,
               data = data)
summary(m_con_relsat)
tidy(m_con_relsat)
```

##### Sensitivity Analyses
###### HC
```{r}
m_con_relsat_sensitivity_hc <- sensemakr(model = m_con_relsat, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_relsat_sensitivity_hc
summary(m_con_relsat_sensitivity_hc)
```


###### Congruency
```{r}
m_con_relsat_sensitivity_con <- sensemakr(model = m_con_relsat, #model
                                treatment = "congruent_contraception_numeric", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_relsat_sensitivity_con
summary(m_con_relsat_sensitivity_con)
```

###### Interaction
```{r}
m_con_relsat_sensitivity_interaction <- sensemakr(model = m_con_relsat, #model
                                treatment = "hc_con_interaction", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_relsat_sensitivity_interaction
summary(m_con_relsat_sensitivity_interaction)
```

#### Controlled Model {.tabset}
##### Model
```{r}
m_con_relsat = lm(relationship_satisfaction ~ contraception_hormonal_numeric +
                 congruent_contraception_numeric + hc_con_interaction +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
               data = data)
summary(m_con_relsat)
tidy(m_con_relsat)
```

##### Sensitivity Analyses
###### HC
```{r}
m_con_relsat_sensitivity_hc <- sensemakr(model = m_con_relsat, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_relsat_sensitivity_hc
summary(m_con_relsat_sensitivity_hc)
```


###### Congruency
```{r}
m_con_relsat_sensitivity_con <- sensemakr(model = m_con_relsat, #model
                                treatment = "congruent_contraception_numeric", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_relsat_sensitivity_con
summary(m_con_relsat_sensitivity_con)
```

###### Interaction
```{r}
m_con_relsat_sensitivity_interaction <- sensemakr(model = m_con_relsat, #model
                                treatment = "hc_con_interaction", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_relsat_sensitivity_interaction
summary(m_con_relsat_sensitivity_interaction)
```

### Sexual Satisfaction {.tabset}
#### Uncontrolled Model {.tabset}
##### Model
```{r}
m_con_sexsat = lm(satisfaction_sexual_intercourse ~ contraception_hormonal_numeric +
                 congruent_contraception_numeric + hc_con_interaction,
               data = data)
summary(m_con_sexsat)
tidy(m_con_sexsat)
```

##### Sensitivity Analyses
###### HC
```{r}
m_con_sexsat_sensitivity_hc <- sensemakr(model = m_con_sexsat, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_sexsat_sensitivity_hc
summary(m_con_sexsat_sensitivity_hc)
```


###### Congruency
```{r}
m_con_sexsat_sensitivity_con <- sensemakr(model = m_con_sexsat, #model
                                treatment = "congruent_contraception_numeric", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_sexsat_sensitivity_con
summary(m_con_sexsat_sensitivity_con)
```

###### Interaction
```{r}
m_con_sexsat_sensitivity_interaction <- sensemakr(model = m_con_sexsat, #model
                                treatment = "hc_con_interaction", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_sexsat_sensitivity_interaction
summary(m_con_sexsat_sensitivity_interaction)
```

#### Controlled Model {.tabset}
##### Model
```{r}
m_con_sexsat = lm(satisfaction_sexual_intercourse ~ contraception_hormonal_numeric +
                 congruent_contraception_numeric + hc_con_interaction +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
               data = data)
summary(m_con_sexsat)
tidy(m_con_sexsat)
```

##### Sensitivity Analyses
###### HC
```{r}
m_con_sexsat_sensitivity_hc <- sensemakr(model = m_con_sexsat, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_sexsat_sensitivity_hc
summary(m_con_sexsat_sensitivity_hc)
```


###### Congruency
```{r}
m_con_sexsat_sensitivity_con <- sensemakr(model = m_con_sexsat, #model
                                treatment = "congruent_contraception_numeric", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_sexsat_sensitivity_con
summary(m_con_sexsat_sensitivity_con)
```

###### Interaction
```{r}
m_con_sexsat_sensitivity_interaction <- sensemakr(model = m_con_sexsat, #model
                                treatment = "hc_con_interaction", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_sexsat_sensitivity_interaction
summary(m_con_sexsat_sensitivity_interaction)
```

### Libido {.tabset}
#### Uncontrolled Model {.tabset}
##### Model
```{r}
m_con_libido = lm(diary_libido_mean ~ contraception_hormonal_numeric +
                 congruent_contraception_numeric + hc_con_interaction,
               data = data)
summary(m_con_libido)
tidy(m_con_libido)
```

##### Sensitivity Analyses
###### HC
```{r}
m_con_libido_sensitivity_hc <- sensemakr(model = m_con_libido, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_libido_sensitivity_hc
summary(m_con_libido_sensitivity_hc)
```


###### Congruency
```{r}
m_con_libido_sensitivity_con <- sensemakr(model = m_con_libido, #model
                                treatment = "congruent_contraception_numeric", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_libido_sensitivity_con
summary(m_con_libido_sensitivity_con)
```

###### Interaction
```{r}
m_con_libido_sensitivity_interaction <- sensemakr(model = m_con_libido, #model
                                treatment = "hc_con_interaction", #predictor
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_libido_sensitivity_interaction
summary(m_con_libido_sensitivity_interaction)
```

#### Controlled Model {.tabset}
##### Model
```{r}
m_con_libido = lm(diary_libido_mean ~ contraception_hormonal_numeric +
                 congruent_contraception_numeric + hc_con_interaction +
                             age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
               data = data)
summary(m_con_libido)
tidy(m_con_libido)
```

##### Sensitivity Analyses
###### HC
```{r}
m_con_libido_sensitivity_hc <- sensemakr(model = m_con_libido, #model
                                treatment = "contraception_hormonal_numeric", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_libido_sensitivity_hc
summary(m_con_libido_sensitivity_hc)
```


###### Congruency
```{r}
m_con_libido_sensitivity_con <- sensemakr(model = m_con_libido, #model
                                treatment = "congruent_contraception_numeric", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_libido_sensitivity_con
summary(m_con_libido_sensitivity_con)
```

###### Interaction
```{r}
m_con_libido_sensitivity_interaction <- sensemakr(model = m_con_libido, #model
                                treatment = "hc_con_interaction", #predictor
                                benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome 
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05, 
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_con_libido_sensitivity_interaction
summary(m_con_libido_sensitivity_interaction)
```





### Sexual Frequency{.tabset}
#### Uncontrolled Model {.tabset}
##### Model
```{r}
m_hc_sexfreqpen = lm(diary_sex_active_sex_mean ~ contraception_hormonal_numeric + congruent_contraception_numeric +
                         hc_con_interaction,
               data = data)
qplot(residuals(m_hc_sexfreqpen))
summary(m_hc_sexfreqpen)
tidy(m_hc_sexfreqpen, conf.int = T)
```

##### Sensitivity Analyses
###### HC
```{r}
m_hc_sexfreqpen_sensitivity_hc <- sensemakr(model = m_hc_sexfreqpen,
                                         treatment = "contraception_hormonal_numeric",
                                        kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05,
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_sexfreqpen_sensitivity_hc
summary(m_hc_sexfreqpen_sensitivity_hc)
```

###### Congruency
```{r}
m_hc_sexfreqpen_sensitivity_con <- sensemakr(model = m_hc_sexfreqpen,
                                         treatment = "congruent_contraception_numeric",
                                        kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05,
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_sexfreqpen_sensitivity_con
summary(m_hc_sexfreqpen_sensitivity_con)
```

###### Interaction
```{r}
m_hc_sexfreqpen_sensitivity_interaction <- sensemakr(model = m_hc_sexfreqpen,
                                         treatment = "hc_con_interaction",
                                        kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05,
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_sexfreqpen_sensitivity_interaction
summary(m_hc_sexfreqpen_sensitivity_interaction)
```

#### Controlled Model {.tabset}
##### Model
```{r}
m_hc_sexfreqpen = lm(diary_sex_active_sex_mean ~ contraception_hormonal_numeric + congruent_contraception_numeric +
                         hc_con_interaction + 
                       age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
               data = data)
qplot(residuals(m_hc_sexfreqpen))
summary(m_hc_sexfreqpen)
tidy(m_hc_sexfreqpen, conf.int = T)
```

##### Sensitivity Analyses
###### HC
```{r}
m_hc_sexfreqpen_sensitivity_hc <- sensemakr(model = m_hc_sexfreqpen,
                                         treatment = "contraception_hormonal_numeric",
                                         benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05,
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_sexfreqpen_sensitivity_hc
summary(m_hc_sexfreqpen_sensitivity_hc)
```

###### Congreuncy
```{r}
m_hc_sexfreqpen_sensitivity_con <- sensemakr(model = m_hc_sexfreqpen,
                                         treatment = "congruent_contraception_numeric",
                                         benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05,
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_sexfreqpen_sensitivity_con
summary(m_hc_sexfreqpen_sensitivity_con)
```

###### Interaction
```{r}
m_hc_sexfreqpen_sensitivity_interaction <- sensemakr(model = m_hc_sexfreqpen,
                                         treatment = "hc_con_interaction",
                                         benchmark_covariates = covariates, #covariates that will be
                                                                   #used to bound the
                                                                   #plausible strength of the
                                                                   #unobserved confounders
                                kd = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the
                                          #treatment
                                ky = 1:3, #these arguments parameterize how many times
                                          #stronger the confounder is related to the outcome
                                q = 1, #fraction of the effect estimate that would have to be
                                       #explained away to be problematic. Setting q = 1,
                                       #means that a reduction of 100% of the current effect
                                       #estimate, that is, a true effect of zero, would be
                                       #deemed problematic.
                                alpha = 0.05,
                                reduce = TRUE #confounder reduce absolute effect size
                                )


m_hc_sexfreqpen_sensitivity_interaction
summary(m_hc_sexfreqpen_sensitivity_interaction)
```

### Masturbation Frequency{.tabset}
#### Uncontrolled Model {.tabset}
##### Model
```{r}
m_hc_masfreq = lm(diary_masturbation_mean ~ contraception_hormonal_numeric + congruent_contraception_numeric +
                       hc_con_interaction,
                     data = data)
qplot(residuals(m_hc_masfreq))
summary(m_hc_masfreq)
tidy(m_hc_masfreq, conf.int = T)
```

##### Sensitivity Analyses
###### HC
```{r}
m_hc_masfreq_sensitivity_hc <- sensemakr(model = m_hc_masfreq,
                                            treatment = "contraception_hormonal_numeric",
                                            kd = 1:3, #these arguments parameterize how many times
                                            #stronger the confounder is related to the
                                            #treatment
                                            ky = 1:3, #these arguments parameterize how many times
                                            #stronger the confounder is related to the outcome
                                            q = 1, #fraction of the effect estimate that would have to be
                                            #explained away to be problematic. Setting q = 1,
                                            #means that a reduction of 100% of the current effect
                                            #estimate, that is, a true effect of zero, would be
                                            #deemed problematic.
                                            alpha = 0.05,
                                            reduce = TRUE #confounder reduce absolute effect size
)


m_hc_masfreq_sensitivity_hc
summary(m_hc_masfreq_sensitivity_hc)
```

###### Congruency
```{r}
m_hc_masfreq_sensitivity_con <- sensemakr(model = m_hc_masfreq,
                                             treatment = "congruent_contraception_numeric",
                                             kd = 1:3, #these arguments parameterize how many times
                                             #stronger the confounder is related to the
                                             #treatment
                                             ky = 1:3, #these arguments parameterize how many times
                                             #stronger the confounder is related to the outcome
                                             q = 1, #fraction of the effect estimate that would have to be
                                             #explained away to be problematic. Setting q = 1,
                                             #means that a reduction of 100% of the current effect
                                             #estimate, that is, a true effect of zero, would be
                                             #deemed problematic.
                                             alpha = 0.05,
                                             reduce = TRUE #confounder reduce absolute effect size
)


m_hc_masfreq_sensitivity_con
summary(m_hc_masfreq_sensitivity_con)
```

###### Interaction
```{r}
m_hc_masfreq_sensitivity_interaction <- sensemakr(model = m_hc_masfreq,
                                                     treatment = "hc_con_interaction",
                                                     kd = 1:3, #these arguments parameterize how many times
                                                     #stronger the confounder is related to the
                                                     #treatment
                                                     ky = 1:3, #these arguments parameterize how many times
                                                     #stronger the confounder is related to the outcome
                                                     q = 1, #fraction of the effect estimate that would have to be
                                                     #explained away to be problematic. Setting q = 1,
                                                     #means that a reduction of 100% of the current effect
                                                     #estimate, that is, a true effect of zero, would be
                                                     #deemed problematic.
                                                     alpha = 0.05,
                                                     reduce = TRUE #confounder reduce absolute effect size
)


m_hc_masfreq_sensitivity_interaction
summary(m_hc_masfreq_sensitivity_interaction)
```

#### Controlled Model {.tabset}
##### Model
```{r}
m_hc_masfreq = lm(diary_masturbation_mean ~ contraception_hormonal_numeric + congruent_contraception_numeric +
                       hc_con_interaction + 
                       age + net_income + relationship_duration_factor +
                       education_years +
                       bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                       religiosity,
                     data = data)
qplot(residuals(m_hc_masfreq))
summary(m_hc_masfreq)
tidy(m_hc_masfreq, conf.int = T)
```

##### Sensitivity Analyses
###### HC
```{r}
m_hc_masfreq_sensitivity_hc <- sensemakr(model = m_hc_masfreq,
                                            treatment = "contraception_hormonal_numeric",
                                            benchmark_covariates = covariates, #covariates that will be
                                            #used to bound the
                                            #plausible strength of the
                                            #unobserved confounders
                                            kd = 1:3, #these arguments parameterize how many times
                                            #stronger the confounder is related to the
                                            #treatment
                                            ky = 1:3, #these arguments parameterize how many times
                                            #stronger the confounder is related to the outcome
                                            q = 1, #fraction of the effect estimate that would have to be
                                            #explained away to be problematic. Setting q = 1,
                                            #means that a reduction of 100% of the current effect
                                            #estimate, that is, a true effect of zero, would be
                                            #deemed problematic.
                                            alpha = 0.05,
                                            reduce = TRUE #confounder reduce absolute effect size
)


m_hc_masfreq_sensitivity_hc
summary(m_hc_masfreq_sensitivity_hc)
```

###### Congreuncy
```{r}
m_hc_masfreq_sensitivity_con <- sensemakr(model = m_hc_masfreq,
                                             treatment = "congruent_contraception_numeric",
                                             benchmark_covariates = covariates, #covariates that will be
                                             #used to bound the
                                             #plausible strength of the
                                             #unobserved confounders
                                             kd = 1:3, #these arguments parameterize how many times
                                             #stronger the confounder is related to the
                                             #treatment
                                             ky = 1:3, #these arguments parameterize how many times
                                             #stronger the confounder is related to the outcome
                                             q = 1, #fraction of the effect estimate that would have to be
                                             #explained away to be problematic. Setting q = 1,
                                             #means that a reduction of 100% of the current effect
                                             #estimate, that is, a true effect of zero, would be
                                             #deemed problematic.
                                             alpha = 0.05,
                                             reduce = TRUE #confounder reduce absolute effect size
)


m_hc_masfreq_sensitivity_con
summary(m_hc_masfreq_sensitivity_con)
```

###### Interaction
```{r}
m_hc_masfreq_sensitivity_interaction <- sensemakr(model = m_hc_masfreq,
                                                     treatment = "hc_con_interaction",
                                                     benchmark_covariates = covariates, #covariates that will be
                                                     #used to bound the
                                                     #plausible strength of the
                                                     #unobserved confounders
                                                     kd = 1:3, #these arguments parameterize how many times
                                                     #stronger the confounder is related to the
                                                     #treatment
                                                     ky = 1:3, #these arguments parameterize how many times
                                                     #stronger the confounder is related to the outcome
                                                     q = 1, #fraction of the effect estimate that would have to be
                                                     #explained away to be problematic. Setting q = 1,
                                                     #means that a reduction of 100% of the current effect
                                                     #estimate, that is, a true effect of zero, would be
                                                     #deemed problematic.
                                                     alpha = 0.05,
                                                     reduce = TRUE #confounder reduce absolute effect size
)


m_hc_masfreq_sensitivity_interaction
summary(m_hc_masfreq_sensitivity_interaction)
```
