==> 18_marginal_effects.Rmd <==
---
title: "Marginal Effects in Controlled Models"
output:
  html_document:
    toc: true
    toc_depth: 4
    toc_float: true
    code_folding: 'show'
    self_contained: false
editor_options: 
  chunk_output_type: console
---
  
  
## Data 
```{r results='hide',message=F,warning=F}
source("0_helpers.R")
library(modelr)
library(tidybayes)

load("data/cleaned_selected_wrangled.rdata")
```


## Sexual Frequency (Penetrative Intercourse) {.tabset}
### Uncontrolled Model
```{r}
m_hc_sexfreqpen = brm(diary_sex_active_sex_sum ~
                        offset(log(number_of_days)) +
                        contraception_hormonal,
                data = data, family = poisson(),
                file = "m_hc_sexfreqpen")
```

#### Data Grid
```{r}
grid = data %>%
  data_grid(number_of_days = 1,
            contraception_hormonal)

```

#### Add Predicted Values
```{r}
data_pred = add_epred_draws(grid, m_hc_sexfreqpen,
                   value = "E[y|A,B]",
                   ndraws = 1000)
```

#### Average Marginal Effect
```{r}
ame = data_pred %>%
  group_by(contraception_hormonal, .draw) %>%  # group by predictors to keep
  summarise(`E[y|A]` = mean(`E[y|A,B]`, na.rm = T)) %>%
  compare_levels(`E[y|A]`, by = contraception_hormonal) %>%  
  rename(`mean difference` = `E[y|A]`) %>%
  mean_qi(`mean difference`, .width = .90)
ame
```

#### Compare Means
```{r}
means = data_pred %>%
  group_by(contraception_hormonal) %>%
  summarise(mean_fit = mean(`E[y|A,B]`, na.rm = T))
print(data.frame(means), digits = 4)
```


### Controlled Model
```{r}
m_hc_sexfreqpen_controlled = brm(diary_sex_active_sex_sum ~
                        offset(log(number_of_days)) +
                        contraception_hormonal +
                        age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
                data = data, family = poisson(),
                file = "m_hc_sexfreqpen_controlled")
```


#### Data Grid
```{r}
grid = data %>%
  data_grid(number_of_days = 1,
            contraception_hormonal,
            age = seq_range(age, n = 5), # measured in full years
            net_income,
            relationship_duration_factor,
            education_years = seq_range(education_years, n = 5), # measured in full years
            bfi_extra = mean(data$bfi_extra),
            bfi_neuro = mean(data$bfi_neuro), 
            bfi_agree = mean(data$bfi_agree), 
            bfi_consc = mean(data$bfi_consc), 
            bfi_open = mean(data$bfi_open),
            religiosity = mean(data$religiosity)
  )

```

#### Add Predicted Values
```{r}
data_pred = add_epred_draws(grid, m_hc_sexfreqpen_controlled,
                   value = "E[y|A,B]",
                   ndraws = 1000)
```

#### Average Marginal Effect
```{r}
ame = data_pred %>%
  group_by(contraception_hormonal, .draw) %>%  # group by predictors to keep
  summarise(`E[y|A]` = mean(`E[y|A,B]`, na.rm = T)) %>%
  compare_levels(`E[y|A]`, by = contraception_hormonal) %>%  
  rename(`mean difference` = `E[y|A]`) %>%
  mean_qi(`mean difference`, .width = .90)
ame
```

#### Compare Means
```{r}
means = data_pred %>%
  group_by(contraception_hormonal) %>%
  summarise(mean_fit = mean(`E[y|A,B]`, na.rm = T))
print(as.data.frame(means), digits = 4)
```


## Masturbation Frequency {.tabset}
### Uncontrolled Model
```{r}
m_hc_masfreq = brm(diary_masturbation_sum ~ offset(log(number_of_days)) +
                     contraception_hormonal,
                data = data, family = poisson(),
                file = "m_hc_masfreq")
```


#### Data Grid
```{r}
grid = data %>%
  data_grid(number_of_days = 1,
            contraception_hormonal)

```

#### Add Predicted Values
```{r}
data_pred = add_epred_draws(grid, m_hc_masfreq,
                   value = "E[y|A,B]",
                   ndraws = 1000)
```

#### Average Marginal Effect
```{r}
ame = data_pred %>%
  group_by(contraception_hormonal, .draw) %>%  # group by predictors to keep
  summarise(`E[y|A]` = mean(`E[y|A,B]`, na.rm = T)) %>%
  compare_levels(`E[y|A]`, by = contraception_hormonal) %>%  
  rename(`mean difference` = `E[y|A]`) %>%
  mean_qi(`mean difference`, .width = .90)
ame
```

#### Compare Means
```{r}
means = data_pred %>%
  group_by(contraception_hormonal) %>%
  summarise(mean_fit = mean(`E[y|A,B]`, na.rm = T))
print(as.data.frame(means), digits = 4)
```

### Controlled Model
```{r}
m_hc_masfreq_controlled = brm(diary_masturbation_sum ~
                        offset(log(number_of_days)) +
                        contraception_hormonal +
                        age + net_income + relationship_duration_factor +
                              education_years +
                              bfi_extra + bfi_neuro + bfi_agree + bfi_consc + bfi_open +
                              religiosity,
                data = data, family = poisson(),
                file = "m_hc_masfreq_controlled")
```


#### Data Grid
```{r}
grid = data %>%
  data_grid(number_of_days = 1,
            contraception_hormonal,
            age = seq_range(age, n = 5), # measured in full years
            net_income,
            relationship_duration_factor,
            education_years = seq_range(education_years, n = 5), # measured in full years
            bfi_extra = mean(data$bfi_extra),
            bfi_neuro = mean(data$bfi_neuro), 
            bfi_agree = mean(data$bfi_agree), 
            bfi_consc = mean(data$bfi_consc), 
            bfi_open = mean(data$bfi_open),
            religiosity = mean(data$religiosity)
  )

```

#### Add Predicted Values
```{r}
data_pred = add_epred_draws(grid, m_hc_masfreq_controlled,
                   value = "E[y|A,B]",
                   ndraws = 1000)
```

#### Average Marginal Effect
```{r}
ame = data_pred %>%
  group_by(contraception_hormonal, .draw) %>%  # group by predictors to keep
  summarise(`E[y|A]` = mean(`E[y|A,B]`, na.rm = T)) %>%
  compare_levels(`E[y|A]`, by = contraception_hormonal) %>%  
  rename(`mean difference` = `E[y|A]`) %>%
  mean_qi(`mean difference`, .width = .90)
ame
```

#### Compare Means
```{r}
means = data_pred %>%
  group_by(contraception_hormonal) %>%
  summarise(mean_fit = mean(`E[y|A,B]`, na.rm = T))
print(as.data.frame(means), digits = 4)
```
